% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_outlier_subsets.R
\name{apply_slice_conditionally}
\alias{apply_slice_conditionally}
\title{Apply slice conditionally}
\usage{
apply_slice_conditionally(x, filter_vars)
}
\arguments{
\item{x}{A tibble, containing the full dataset which will be subset with \code{\link[=slice_conditionally]{slice_conditionally()}} and bound back to the original dataset \code{x}}

\item{filter_vars}{A list of quosures to be used in \code{\link[dplyr:filter]{dplyr::filter()}} to subset \code{y}}

\item{n_min}{integer, the number of bottom outliers to remove}

\item{n_max}{integer, the number of top outliers to remove}
}
\description{
Apply slice conditionally
}
\details{
the tibble \code{x} must contain the columns \code{data}, \code{outcome_colname}, \code{n_min}, and \code{n_max}
}
\keyword{internal}
