% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_level_mod_calcs.R
\name{apportion_heterogeneity_ml}
\alias{apportion_heterogeneity_ml}
\title{Apportion heterogeneity of a multi-level meta-analytic model}
\usage{
apportion_heterogeneity_ml(fitted_model)
}
\value{
A named numeric vector of length 2.
}
\description{
Estimates how much the total variance \eqn{I^2} can be attributed to between- and within- cluster heterogeneity separately.
}
\details{
From \url{http://www.metafor-project.org/doku.php/tips:i2_multilevel_multivariate}
}
\seealso{
Other Meta-analysis statistics:
\code{\link[=calc_I2_ml]{calc_I2_ml()}},
\code{\link[=compare_ml_MA]{compare_ml_MA()}}
}
\concept{Meta-analysis statistics}
