% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_transformation_type.R
\name{assign_transformation_type}
\alias{assign_transformation_type}
\title{Assign back-transformation type to be applied to analysis point-estimates}
\usage{
assign_transformation_type(
  response_transformation = character(1L),
  link_fun = character(1L)
)
}
\arguments{
\item{response_transformation}{Character vector of length 1L containing the analysis response transformation}

\item{link_fun}{Character vector of length 1L containing the analysis link function}
}
\value{
A character vector of length 1L
}
\description{
Assign back-transformation type to be applied to analysis point-estimates
}
