% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_prediction_data.R
\name{augment_prediction_data}
\alias{augment_prediction_data}
\title{Augment analyst out-of-sample prediction data according to the outcome of pointblank interrogation}
\usage{
augment_prediction_data(.data, checks, dataset)
}
\arguments{
\item{.data}{A data.frame of out of-sample-predictions analyst submission data}

\item{checks}{A data.frame of \verb{pointblank::} interrogation output for \code{.data}}

\item{dataset}{A character string equal to either "blue tit" or "eucalyptus"}
}
\value{
A dataframe of individual analyst submission containing augmented
out-of-sample prediction data ready for meta-analysis.
}
\description{
\code{augment_prediction_data()} will relabel and remove extraneous columns from
out-of-sample prediction data.
}
\details{
Where there are missing variables in \code{.data} or \code{checks} is missing,
\code{NA} will be returned for the augmented prediction data instead.
}
\examples{
# for testing and dev purposes:
# safe_augment <- purrr:::safely(augment_prediction_data)
}
