% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_descriptive_statistics.R
\name{calc_analyses_per_team}
\alias{calc_analyses_per_team}
\title{Calculate total number of analyses per team for a given subset}
\usage{
calc_analyses_per_team(data, subset_name = character(1L))
}
\arguments{
\item{data}{A tibble containing the data to be analysed.}

\item{subset_name}{A character vector of length 1, the name of the subset of data being analysed.}
}
\value{
A tibble containing the number of \code{analyses} per \code{team} for each \code{dataset} in a given \code{subset_name}.
}
\description{
Calculates the number of analyses conducted by each team for each dataset in a given subset.
}
\examples{
ManyEcoEvo::ManyEcoEvo \%>\% 
select(data) \%>\% 
  unnest(everything()) \%>\% 
  prepare_df_for_summarising() \%>\% 
  calc_analyses_per_team("All")
}
