% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_level_mod_calcs.R
\name{calc_I2_ml}
\alias{calc_I2_ml}
\title{Calculate I2 for a multilevel meta-analytic model}
\usage{
calc_I2_ml(fitted_model)
}
\arguments{
\item{fitted_model}{of class rma.mv}
}
\value{
A numeric vector of length 1.
}
\description{
Calculate I2 for a multilevel meta-analytic model
}
\details{
from \href{http://www.metafor-project.org/doku.php/tips:i2_multilevel_multivariate]}{http://www.metafor-project.org/doku.php/tips:i2_multilevel_multivariate} described by Nakagawa & Santos, 2012.
}
\seealso{
Other Meta-analysis statistics:
\code{\link[=apportion_heterogeneity_ml]{apportion_heterogeneity_ml()}},
\code{\link[=compare_ml_MA]{compare_ml_MA()}}
}
\concept{Meta-analysis statistics}
