% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_descriptive_statistics.R
\name{calc_summary_stats_binary}
\alias{calc_summary_stats_binary}
\title{Calculate summary statistics for binary summary variables}
\usage{
calc_summary_stats_binary(data, subset_name = character(1L))
}
\arguments{
\item{data}{A tibble containing the data to be analysed.}

\item{subset_name}{A character vector of length 1, the name of the subset of data being analysed.}
}
\value{
A tibble containing the sum of binary variables used in analyses of each dataset for a given subset.
}
\description{
Calculates the total number of analyses using linear models, mixed models,
and Bayesian models for each dataset, for a given subset.
See \code{\link[=prepare_df_for_summarising]{prepare_df_for_summarising()}} for details on the binary variables.
}
\examples{
ManyEcoEvo::ManyEcoEvo \%>\%
select(data) \%>\%
unnest(everything()) \%>\%
prepare_df_for_summarising() \%>\%
calc_summary_stats_binary("All")
}
