% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_descriptive_statistics.R
\name{calc_summary_stats_numeric}
\alias{calc_summary_stats_numeric}
\title{Calculate summary statistics for numeric summary variables}
\usage{
calc_summary_stats_numeric(data, subset_name = character(1L))
}
\arguments{
\item{data}{A tibble containing the data to be analysed.}

\item{subset_name}{A character vector of length 1, the name of the subset of
data being analysed.}
}
\value{
A tibble containing the mean, standard deviation, minimum and maximum
values for each numeric variable used in analyses of each dataset for a given subset.
}
\description{
Calculates the mean, standard deviation, minimum and maximum for each numeric summary variable (See \code{\link[=prepare_df_for_summarising]{prepare_df_for_summarising()}}).

for numeric
variables used in analyses of each dataset, for a given subset.
Summary statistics are aggregated across variable type \(Number of fixed variables
within the analysis, number of random variables within the analysis,
analysis samplesize, number of interaction terms within the analysis\).
}
\examples{
ManyEcoEvo::ManyEcoEvo \%>\% 
select(data) \%>\% 
  unnest(everything()) \%>\% 
  prepare_df_for_summarising() \%>\% 
  calc_summary_stats_numeric("All")
}
