% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_descriptive_statistics.R
\name{calc_teams_per_dataset}
\alias{calc_teams_per_dataset}
\title{Calculate the number of teams per dataset for a given subset}
\usage{
calc_teams_per_dataset(data, subset_name = character(1L))
}
\arguments{
\item{data}{A tibble containing the data to be analysed.}

\item{subset_name}{A character vector of length 1, the name of the subset of data being analysed.}
}
\value{
A tibble containing the number of \code{teams} per \code{dataset} for a given \code{subset_name}.
}
\description{
Calculate the number of teams per dataset for a given subset
}
\examples{
ManyEcoEvo::ManyEcoEvo \%>\% 
select(data) \%>\% 
unnest(everything()) \%>\% 
prepare_df_for_summarising() \%>\% 
calc_teams_per_dataset("all")
}
