% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_deviation_score.R
\name{calculate_deviation_score}
\alias{calculate_deviation_score}
\title{Calculate deviation from meta-analytic mean}
\usage{
calculate_deviation_score(data, meta_analytic_model)
}
\arguments{
\item{data}{Dataset containing the column \code{"Z"} \(standardised out of sample prediction piont estimates\) or standardised effect-sizes, \code{"Zr"}.}

\item{meta_analytic_model}{Fitted meta-analytic model of class \code{rma}}
}
\value{
\code{data} but with an additional column of deviation score estimates \code{"abs_deviation_score_estimate"}
}
\description{
Calculate the absolute deviation of the standardised effect size from the meta-analytic mean
}
\examples{
  # Example Usage: TODO - this is incorrect bc box-cox transformation is included in meta_analyse_datasets_wrapper
# library(tidyverse); library(targets)
# tar_load(meta_analysis_outputs)
# calculate_deviation_score(meta_analysis_outputs$data[[1]],
#                           meta_analysis_outputs$MA_mod[[1]])
}
