% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sorensen_diversity_index.R
\name{calculate_sorensen_diversity_index}
\alias{calculate_sorensen_diversity_index}
\title{Calculate mean sorensen diversity index values for an analyst dataset}
\usage{
calculate_sorensen_diversity_index(.data, .id = character())
}
\arguments{
\item{.data}{A variable diversity dataset with columns for each unique variable used in all analyses. Values are \code{NA} if the analysis did not use that variable, and take the value of the variable as a character string if the analysis did use that variable. Each unique analysis is a row.}

\item{.id}{character string of the analysis identifier column}
}
\value{
A tibble containing the variables \code{id_col}, \code{mean_diversity_index} and \code{num_variables}
}
\description{
Calculate mean sorensen diversity index values for an analyst dataset
}
