% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_MA_inputs.R
\name{compute_MA_inputs}
\alias{compute_MA_inputs}
\title{Compute meta-analysis inputs for a nested-dataframe containing different datasets/subsets of analyst data}
\usage{
compute_MA_inputs(ManyEcoEvo, estimate_type = NULL)
}
\arguments{
\item{ManyEcoEvo}{A dataframe grouped by the character columns \code{dataset}, \code{estimate_type}, \code{exclusion_set}. Each group corresponds to a subset of the full \code{dataset}, and has the subset analyst data  stored in \code{data}, with its corresponding subset \code{diversity_data}.}

\item{estimate_type}{character string, one of "Zr", "yi", "y25", "y50", "y75".}
}
\value{
A dataframe that includes the additional columns in \code{ManyEcoEvo}, but with added columns \code{diversity_indices} and \code{effects_analysis}.
}
\description{
Computes the Sorensen diversity indices and joins it to data in preparation for meta-analysing all subsets of data with \code{\link[=meta_analyse_datasets]{meta_analyse_datasets()}}.
}
\details{
Computes Sorensen diversity indices \code{diversity_indices} for each subset of data returning them in the list-column \code{diversity_indices} and joins them to the relevant subset of processed analyst \code{data} within the list-column \code{effects_analysis}.

Note that, the name of the subset is derived from the functions within by \code{\link[=subset_fns_yi]{subset_fns_yi()}} and/or \code{\link[=subset_fns_Zr]{subset_fns_Zr()}} called in the previous step of data processing \code{\link[=generate_exclusion_subsets]{generate_exclusion_subsets()}}.
Should the user wish to skip the \code{generate_exclusion_subsets()} step, they can supply arbitrary values for \code{exclusion_set} and the function will still work.

Should be run after \emph{all} data subsetting is complete, otherwise the diversity indices will need to be recalculated.
}
\seealso{
\code{\link[=apply_sorensen_calc]{apply_sorensen_calc()}} is used to calculate the Sorensen diversity indices.

Other Multi-dataset Wrapper Functions:
\code{\link[=apply_VZ_exclusions]{apply_VZ_exclusions()}},
\code{\link[=generate_exclusion_subsets]{generate_exclusion_subsets()}},
\code{\link[=generate_outlier_subsets]{generate_outlier_subsets()}},
\code{\link[=generate_rating_subsets]{generate_rating_subsets()}},
\code{\link[=generate_yi_subsets]{generate_yi_subsets()}},
\code{\link[=make_viz]{make_viz()}},
\code{\link[=meta_analyse_datasets]{meta_analyse_datasets()}},
\code{\link[=prepare_ManyEcoEvo]{prepare_ManyEcoEvo()}},
\code{\link[=prepare_ManyEcoEvo_yi]{prepare_ManyEcoEvo_yi()}},
\code{\link[=prepare_response_variables]{prepare_response_variables()}},
\code{\link[=prepare_response_variables_yi]{prepare_response_variables_yi()}},
\code{\link[=summarise_analysis_types]{summarise_analysis_types()}},
\code{\link[=summarise_conclusions]{summarise_conclusions()}},
\code{\link[=summarise_model_composition]{summarise_model_composition()}},
\code{\link[=summarise_reviews]{summarise_reviews()}},
\code{\link[=summarise_sorensen_index]{summarise_sorensen_index()}},
\code{\link[=summarise_variable_counts]{summarise_variable_counts()}}
}
\concept{Multi-dataset Wrapper Functions}
