% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metaanalysis_inputs.R
\name{compute_metaanalysis_inputs}
\alias{compute_metaanalysis_inputs}
\title{Compute all metaanalysis inputs for different types of estimates}
\usage{
compute_metaanalysis_inputs(.data, estimate_type = character(1L))
}
\arguments{
\item{.data}{A dataset containing all necessary values for computing meta-analysis inputs}

\item{estimate_type}{character string, one of "Zr", "yi", "y25", "y50", "y75".}
}
\value{
A dataframe with processed data ready for meta-analysis
}
\description{
Compute all metaanalysis inputs for different types of estimates
}
