% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{conversion}
\alias{conversion}
\title{Apply back-transformation to beta estimates}
\usage{
conversion(beta, se, transformation, sim = 10000)
}
\arguments{
\item{beta}{Beta estimate, numeric vector of length 1.}

\item{se}{Standard error of the \code{beta} estimate, numeric vector of length 1}

\item{transformation}{Character string describing transformation}

\item{sim}{Number of simulations to use during back-transformation. Defaults to $10000$.}
}
\value{
The outputs of a back-transformation function, see family back-transformations
}
\description{
Conditionally apply back-transformation functions depending on the value of \code{transformation}.
}
\details{
\code{transformation} character strings may take the values:
\itemize{
\item "log"
\item "logit"
\item "probit"
\item "square"
\item "probit"
\item "double_transformation"
\item "cube"
\item "inverse"
\item "square_root"
\item "powerX", where \code{X} is a numeric
\item "divided.by.X", where \code{X} is a numeric
}
}
\seealso{
\code{\link[=conversion_2]{conversion_2()}}, \code{\link[=back]{back()}}

Other Back-transformation:
\code{\link{back}},
\code{\link[=back_transform_response_vars_yi]{back_transform_response_vars_yi()}},
\code{\link[=clean_response_transformation]{clean_response_transformation()}},
\code{\link[=conversion_2]{conversion_2()}},
\code{\link[=convert_predictions]{convert_predictions()}}
}
\concept{Back-transformation}
