% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_2.R
\name{conversion_2}
\alias{conversion_2}
\title{Conditionally apply back-transformation}
\usage{
conversion_2(beta, se, response_transformation, link_fun, sim = 10000)
}
\arguments{
\item{beta}{Beta estimate, numeric vector of length 1.}

\item{se}{Standard error of the \code{beta} estimate, numeric vector of length 1}

\item{response_transformation}{Character string describing transformation}

\item{link_fun}{Character string describing link function}

\item{sim}{Number of simulations to use during back-transformation. Defaults to \code{10000}.}
}
\value{
The outputs of a back-transformation function, see family back-transformations
}
\description{
Conditionally apply back-transformation functions depending on the value of \code{transformation}
}
\seealso{
\code{\link[=conversion]{conversion()}}

Other Back-transformation:
\code{\link{back}},
\code{\link[=back_transform_response_vars_yi]{back_transform_response_vars_yi()}},
\code{\link[=clean_response_transformation]{clean_response_transformation()}},
\code{\link[=conversion]{conversion()}},
\code{\link[=convert_predictions]{convert_predictions()}}
}
\concept{Back-transformation}
