% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_variable_counts.R
\name{count_analyses_variables_used}
\alias{count_analyses_variables_used}
\title{Count number of analyses each variable is used}
\usage{
count_analyses_variables_used(data)
}
\arguments{
\item{data}{A dataframe containing the variables \code{dataset}, \code{id_col} and columns filled with the name of the variable indicating variable usage.}
}
\value{
A dataframe of counts \code{n} each \code{variable} is used across all analyses within a given \code{dataset}
}
\description{
Count number of analyses each variable is used
}
\details{
Values in variable columns are character strings or \code{NA}. \code{NA} values are treated as absence of variable in a given analysis (observation / row), while character strings are treated as presence of that variable in a given analysis.

Variable presence in an analysis is converted to numeric 1's/0's and then summed to calculate total times the analysis is used across all analyses (\code{n}).
}
\examples{
library(ManyEcoEvo)
ManyEcoEvo \%>\%
  select(diversity_data) \%>\%
  tidyr::unnest(diversity_data) \%>\%
  count_analyses_variables_used()
}
\author{
Hannah S. Fraser

Elliot Gould
}
