% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_analysis_types.R
\name{count_binary_coded_features}
\alias{count_binary_coded_features}
\title{Summarise binary coded features of analyses}
\usage{
count_binary_coded_features(data)
}
\arguments{
\item{data}{A dataframe containing the variables \code{dataset}, and \code{lm}, \code{mixed_model}, \code{Bayesian}, which are coded as binary numeric vectors.}
}
\value{
A dataframe with the variables
}
\description{
Summarise binary coded features of analyses
}
\examples{
ManyEcoEvo \%>\%
  filter(dataset == "eucalyptus") \%>\%
  ungroup() \%>\%
  select(data) \%>\%
  unnest(data) \%>\%
  mutate(
    lm = ifelse(linear_model == "linear", 1, 0), # TODO move into master processing
    glm = ifelse(linear_model == "generalised", 1, 0),
    Bayesian = as.numeric(Bayesian)
  ) \%>\%
  count_binary_coded_features()
}
\author{
Hannah S. Fraser

Elliot Gould
}
