% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_descriptive_statistics.R
\name{count_conclusions}
\alias{count_conclusions}
\title{Count the number of different conclusions made by analysts across each dataset.}
\usage{
count_conclusions(data, subset_name = character(1L))
}
\arguments{
\item{data}{A tibble containing the data to be analysed.}

\item{subset_name}{A character vector of length 1, the name of the subset of data being analysed.}
}
\value{
A tibble containing counts of each conclusion type made by analysts across each dataset, for a given subset.
}
\description{
Takes the first analysis per team per dataset, not each analysis that was submitted.
Thus filters for analyses where split_id == 1 and where analysis_id == 1
}
