% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_analysis_types.R
\name{count_teams_analyses}
\alias{count_teams_analyses}
\title{Summarise number of analyst teams and total analyses per dataset}
\usage{
count_teams_analyses(data)
}
\arguments{
\item{data}{A dataframe containing the variables \code{TeamIdentifier} and \code{dataset}}
}
\value{
A dataframe with the columns  \code{dataset}, \code{total_teams} and \code{total_analyses} equal in number of rows to the number of unique values within the \code{dataset} variable of the input \code{data}.
}
\description{
Summarise number of analyst teams and total analyses per dataset
}
\examples{
ManyEcoEvo \%>\%
  filter(dataset == "blue tit") \%>\%
  ungroup() \%>\%
  select(data) \%>\%
  unnest(data) \%>\%
  count_teams_analyses()
}
\author{
Hannah S. Fraser

Elliot Gould
}
