% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/back_transformations.R
\name{divide_back}
\alias{divide_back}
\title{Back transform beta estimates or out-of-sample predictions from models whose response variable has been divided by some number}
\usage{
divide_back(beta, se, sim, n)
}
\arguments{
\item{beta}{Analyst beta estimate. Numeric vector of length 1.}

\item{se}{Standard error of analyst's beta estimate. Numeric vector of length 1.}

\item{sim}{Number of simulations. Numeric vector of length 1.}

\item{n}{Numeric vector of length 1 describing the value of the divisor.}
}
\value{
data frame containing the mean estimate, its standard error, and quantiles
}
\description{
Back transform beta estimates or out-of-sample predictions from models whose response variable has been divided by some number
}
\seealso{
Other back transformation: 
\code{\link{conversion_2}()},
\code{\link{conversion}()},
\code{\link{cube_back}()},
\code{\link{identity_back}()},
\code{\link{inverse_back}()},
\code{\link{log_back}()},
\code{\link{logit_back}()},
\code{\link{power_back}()},
\code{\link{probit_back}()},
\code{\link{square_back}()},
\code{\link{square_root_back}()}
}
\concept{back transformation}
