% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_to_Zr.R
\name{est_to_zr}
\alias{est_to_zr}
\title{Convert estimate to Zr}
\usage{
est_to_zr(beta_estimate, beta_SE, adjusted_df)
}
\arguments{
\item{beta_estimate}{numeric vector, beta estimate of an analysis of either blue tit or eucalyptus dataset}

\item{beta_SE}{numeric vector, standard error of the \code{beta_estimate}}

\item{adjusted_df}{numeric vector of the adjusted degrees of freedom for the analysis generating \code{beta_SE} and \code{beta_estimate}}
}
\value{
A named list of length 2 containing converted Zr and VZr values
}
\description{
Convert estimate to Zr
}
\details{
Convert beta estimate, standard error and degrees of freedom into \code{r} and then into \code{Zr} and its sampling variance
}
\seealso{
Other Analysis-level functions:
\code{\link[=Z_VZ_preds]{Z_VZ_preds()}},
\code{\link[=apply_sorensen_calc]{apply_sorensen_calc()}},
\code{\link[=back_transform_response_vars_yi]{back_transform_response_vars_yi()}},
\code{\link[=box_cox_transform]{box_cox_transform()}},
\code{\link[=convert_predictions]{convert_predictions()}},
\code{\link[=folded_params]{folded_params()}},
\code{\link[=pred_to_Z]{pred_to_Z()}},
\code{\link[=variance_box_cox]{variance_box_cox()}}
}
\concept{Analysis-level functions}
