% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_boxcox_ratings_cont.R
\name{fit_boxcox_ratings_cont}
\alias{fit_boxcox_ratings_cont}
\title{Fit model of boxcox deviation scores as function of continuous ratings}
\usage{
fit_boxcox_ratings_cont(
  data,
  outcome,
  outcome_var,
  ...,
  env = rlang::caller_env()
)
}
\arguments{
\item{data}{Data for model fitting}

\item{outcome}{outcome variable, unquoted.}

\item{outcome_var}{Variance of the \code{outcome} variable}
}
\value{
An object of class \code{lme4::lmerMod-class}
}
\description{
Fit an lmer model of the box-cox transformed deviation from the meta-analytic mean scores as a function of continuous peer-review ratings
}
\details{
The model is fitted using the \code{\link[lme4:lmer]{lme4::lmer()}} function. The outcome variable is selected by the user \code{outcome}, and the variance of the outcome variable is specified in the argument \code{outcome_var}. The model is fitted with the continuous ratings predictor \code{RateAnalysis} and a random effect for \code{ReviewerId}.

Note that the variables \code{study_id}, \code{outcome}, \code{outcome_var}, and the list-column \code{review_data} must be present in \code{data}. The \code{review_data} column is unnested (see \code{\link[tidyr:unnest]{tidyr::unnest()}}) by the function before fitting the model. Within \code{review_data}, the variables \code{RateAnalysis} and \code{ReviewerId} must be present.
}
\seealso{
Other Model fitting and meta-analysis:
\code{\link[=fit_MA_mv]{fit_MA_mv()}},
\code{\link[=fit_boxcox_ratings_cat]{fit_boxcox_ratings_cat()}},
\code{\link[=fit_boxcox_ratings_ord]{fit_boxcox_ratings_ord()}},
\code{\link[=fit_metafor_mv]{fit_metafor_mv()}},
\code{\link[=fit_metafor_mv_reduced]{fit_metafor_mv_reduced()}},
\code{\link[=fit_metafor_uni]{fit_metafor_uni()}},
\code{\link[=fit_multivar_MA]{fit_multivar_MA()}},
\code{\link[=fit_sorensen_glm]{fit_sorensen_glm()}},
\code{\link[=fit_uni_mixed_effects]{fit_uni_mixed_effects()}},
\code{\link[=poss_fit_boxcox_ratings_cat]{poss_fit_boxcox_ratings_cat()}},
\code{\link[=poss_fit_boxcox_ratings_cont]{poss_fit_boxcox_ratings_cont()}},
\code{\link[=poss_fit_boxcox_ratings_ord]{poss_fit_boxcox_ratings_ord()}},
\code{\link[=poss_fit_metafor_mv]{poss_fit_metafor_mv()}},
\code{\link[=poss_fit_uni_mixed_effects]{poss_fit_uni_mixed_effects()}}
}
\concept{Model fitting and meta-analysis}
