% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_metafor_mv.R
\name{fit_metafor_mv_reduced}
\alias{fit_metafor_mv_reduced}
\title{Fit reduced metaregression model}
\usage{
fit_metafor_mv_reduced(estimate, variance, estimate_type = character(1L), data)
}
\arguments{
\item{estimate}{Numeric vector}

\item{variance}{Numeric vector}

\item{estimate_type}{Character vector of either "Zr", "y50", "y25", "y75".}

\item{data}{Dataframe containing estimates and variances with case id column \code{study_id}.}
}
\value{
Object of class \code{rma.mv}
}
\description{
Fit reduced metaregression model
}
\seealso{
Other Model fitting and meta-analysis:
\code{\link[=fit_MA_mv]{fit_MA_mv()}},
\code{\link[=fit_boxcox_ratings_cat]{fit_boxcox_ratings_cat()}},
\code{\link[=fit_boxcox_ratings_cont]{fit_boxcox_ratings_cont()}},
\code{\link[=fit_boxcox_ratings_ord]{fit_boxcox_ratings_ord()}},
\code{\link[=fit_metafor_mv]{fit_metafor_mv()}},
\code{\link[=fit_metafor_uni]{fit_metafor_uni()}},
\code{\link[=fit_multivar_MA]{fit_multivar_MA()}},
\code{\link[=fit_sorensen_glm]{fit_sorensen_glm()}},
\code{\link[=fit_uni_mixed_effects]{fit_uni_mixed_effects()}},
\code{\link[=poss_fit_boxcox_ratings_cat]{poss_fit_boxcox_ratings_cat()}},
\code{\link[=poss_fit_boxcox_ratings_cont]{poss_fit_boxcox_ratings_cont()}},
\code{\link[=poss_fit_boxcox_ratings_ord]{poss_fit_boxcox_ratings_ord()}},
\code{\link[=poss_fit_metafor_mv]{poss_fit_metafor_mv()}},
\code{\link[=poss_fit_uni_mixed_effects]{poss_fit_uni_mixed_effects()}}
}
\concept{Model fitting and meta-analysis}
