% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt_multivar_MA.R
\name{fit_multivar_MA}
\alias{fit_multivar_MA}
\title{Fit a multivariate meta-regression model}
\usage{
fit_multivar_MA(data_tbl, N = 5, ..., env = rlang::caller_env())
}
\arguments{
\item{data_tbl}{Data for model fitting}

\item{N}{threshold for the number of analyses that must have been conducted using mixed effects models to include the binary predictor \code{mixed_model} in the meta-regression. Defaults to 5.}

\item{...}{Additional arguments passed to \link[lme4:lmer]{lme4::lmer}}

\item{env}{Environment in which to evaluate the formula, defaults to the calling environment}
}
\value{
An object of class lmer.
}
\description{
Fit a multivariate meta-regression model that models the effect of peer-review ratings on the deviation from the meta-analytic mean (both continuous and categorical ratings), mean Sorensen's index, and/or whether the analysis uses a mixed effects model, or not.
}
\details{
Depending on whether enough analyses in \code{data_tbl} have been conducted with the \code{mixed_model} variable, the function will fit a model with or without the predictor \code{mixed_model}.

Expects the following columns in \code{data_tbl}:
\itemize{
\item \code{RateAnalysis}: continuous peer-review ratings
\item \code{PublishableAsIs}: categorical peer-review ratings
\item \code{mean_diversity_index}: mean Sorensen's index
\item \code{box_cox_abs_deviation_score_estimate}: response variable, Box-Cox transformed deviation from the meta-analytic mean effect-size for each analysis
\item \code{mixed_model}: binary variable indicating whether the analysis used a mixed effects model or not
\item \code{ReviewerId}: reviewer identifier
\item one of \code{study_id} or \code{id_col} to uniquely identify each analysis for checking that the threshold \code{N} is met.
}
}
\seealso{
Other Model fitting and meta-analysis:
\code{\link[=fit_MA_mv]{fit_MA_mv()}},
\code{\link[=fit_boxcox_ratings_cat]{fit_boxcox_ratings_cat()}},
\code{\link[=fit_boxcox_ratings_cont]{fit_boxcox_ratings_cont()}},
\code{\link[=fit_boxcox_ratings_ord]{fit_boxcox_ratings_ord()}},
\code{\link[=fit_metafor_mv]{fit_metafor_mv()}},
\code{\link[=fit_metafor_mv_reduced]{fit_metafor_mv_reduced()}},
\code{\link[=fit_metafor_uni]{fit_metafor_uni()}},
\code{\link[=fit_sorensen_glm]{fit_sorensen_glm()}},
\code{\link[=fit_uni_mixed_effects]{fit_uni_mixed_effects()}},
\code{\link[=poss_fit_boxcox_ratings_cat]{poss_fit_boxcox_ratings_cat()}},
\code{\link[=poss_fit_boxcox_ratings_cont]{poss_fit_boxcox_ratings_cont()}},
\code{\link[=poss_fit_boxcox_ratings_ord]{poss_fit_boxcox_ratings_ord()}},
\code{\link[=poss_fit_metafor_mv]{poss_fit_metafor_mv()}},
\code{\link[=poss_fit_uni_mixed_effects]{poss_fit_uni_mixed_effects()}}
}
\concept{Model fitting and meta-analysis}
