% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sorensen_glm.R
\name{fit_sorensen_glm}
\alias{fit_sorensen_glm}
\title{Fit univariate glm of deviation scores on sorensen diversity index}
\usage{
fit_sorensen_glm(data)
}
\arguments{
\item{data}{A \code{data.frame} containing Box-Cox transformed absolute deviation from the meta-analytic mean scores and the mean Sorensen's scores for each analysis}
}
\value{
A fitted model object of class \code{glm} and \code{parsnip}
}
\description{
Fit univariate glm of deviation scores on sorensen diversity index
}
\seealso{
Other Model fitting and meta-analysis:
\code{\link[=fit_MA_mv]{fit_MA_mv()}},
\code{\link[=fit_boxcox_ratings_cat]{fit_boxcox_ratings_cat()}},
\code{\link[=fit_boxcox_ratings_cont]{fit_boxcox_ratings_cont()}},
\code{\link[=fit_boxcox_ratings_ord]{fit_boxcox_ratings_ord()}},
\code{\link[=fit_metafor_mv]{fit_metafor_mv()}},
\code{\link[=fit_metafor_mv_reduced]{fit_metafor_mv_reduced()}},
\code{\link[=fit_metafor_uni]{fit_metafor_uni()}},
\code{\link[=fit_multivar_MA]{fit_multivar_MA()}},
\code{\link[=fit_uni_mixed_effects]{fit_uni_mixed_effects()}},
\code{\link[=poss_fit_boxcox_ratings_cat]{poss_fit_boxcox_ratings_cat()}},
\code{\link[=poss_fit_boxcox_ratings_cont]{poss_fit_boxcox_ratings_cont()}},
\code{\link[=poss_fit_boxcox_ratings_ord]{poss_fit_boxcox_ratings_ord()}},
\code{\link[=poss_fit_metafor_mv]{poss_fit_metafor_mv()}},
\code{\link[=poss_fit_uni_mixed_effects]{poss_fit_uni_mixed_effects()}}
}
\concept{Model fitting and meta-analysis}
