% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_cox_transform.R
\name{folded_params}
\alias{folded_params}
\title{Calculate the folded parameters for the Box-Cox transformation}
\usage{
folded_params(abs_dev_score, VZr)
}
\arguments{
\item{abs_dev_score}{The absolute deviation score}

\item{VZr}{The variance of the standardised effect size}
}
\value{
A named list containing the mean \code{fold_mu} and variance \code{fold_v} of the folded parameters
}
\description{
Calculate the folded parameters for the Box-Cox transformation
}
\seealso{
Other Box-Cox transformation:
\code{\link[=box_cox_transform]{box_cox_transform()}},
\code{\link[=variance_box_cox]{variance_box_cox()}}

Other Analysis-level functions:
\code{\link[=Z_VZ_preds]{Z_VZ_preds()}},
\code{\link[=apply_sorensen_calc]{apply_sorensen_calc()}},
\code{\link[=back_transform_response_vars_yi]{back_transform_response_vars_yi()}},
\code{\link[=box_cox_transform]{box_cox_transform()}},
\code{\link[=convert_predictions]{convert_predictions()}},
\code{\link[=est_to_zr]{est_to_zr()}},
\code{\link[=pred_to_Z]{pred_to_Z()}},
\code{\link[=variance_box_cox]{variance_box_cox()}}
}
\concept{Analysis-level functions}
\concept{Box-Cox transformation}
