% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_outlier_subsets.R
\name{formulae_match}
\alias{formulae_match}
\title{Match formulae to outcome variables}
\usage{
formulae_match(x, y)
}
\arguments{
\item{x}{A character vector}

\item{y}{A vector}
}
\value{
A named list of formulae
}
\description{
Match formulae to outcome variables
}
\details{
This function is used to match formulae to variables with a \code{data.frame}. If \code{x} and \code{y} are the same vector, then the formulae will be matched to the same variable.
}
\keyword{internal}
