% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_collinearity_subset.R
\name{generate_collinearity_subset}
\alias{generate_collinearity_subset}
\title{Generate Collinearity Data Subset}
\usage{
generate_collinearity_subset(ManyEcoEvo, collinearity_subset)
}
\arguments{
\item{ManyEcoEvo}{a ManyEcoEvo dataframe containing formatted raw \code{data}, formatted \code{diversity_data}, the \code{estimate_type},  \code{dataset}, \code{publishable_subset}, and \code{exclusion_set}. See details.}

\item{collinearity_subset}{a dataframe containing the column \code{response_id} containing response ID's to be included in the expert subset}
}
\value{
A ManyEcoEvo dataframe with added column \code{expertise_subset} with new subsets of \code{data} and \code{diversity_data}
}
\description{
This function generates a subset of the data that is used to demonstrate the
effects of collinearity on regression models. The data is generated by
sampling from a multivariate normal distribution with a specified correlation
matrix.
}
\details{
#'

Note that this function needs to be run on \code{ManyEcoEvo} after the following functions have been run (See examples):
\itemize{
\item \code{prepare_response_variables()}
\item \code{generate_exclusion_subsets()}
\item \code{generate_rating_subsets()}
}

\code{generate_collinearity_subset()} only creates expertise subsets based on the full dataset where \code{exclusion_set == "complete"} and \code{publishable_subset == "All"} and \code{expertise_subset == "All"}.
}
\examples{
ManyEcoEvo \%>\%
prepare_response_variables(estimate_type = "Zr") |>
generate_exclusion_subsets(estimate_type = "Zr") |>
generate_rating_subsets() |>
generate_expertise_subsets(ManyEcoEvo:::expert_subset) |>
generate_collinearity_subset(collinearity_subset = collinearity_subset)
}
