% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_expertise_subsets.R
\name{generate_expertise_subsets}
\alias{generate_expertise_subsets}
\title{Generate Expertise Data Subsets}
\usage{
generate_expertise_subsets(ManyEcoEvo, expert_subset)
}
\arguments{
\item{ManyEcoEvo}{a ManyEcoEvo dataframe containing formatted raw \code{data}, formatted \code{diversity_data}, the \code{estimate_type},  \code{dataset}, \code{publishable_subset}, and \code{exclusion_set}. See details.}

\item{expert_subset}{a dataframe containing the column \code{response_id} containing response ID's to be included in the expert subset}
}
\value{
A ManyEcoEvo dataframe with added column \code{expertise_subset} with new subsets of \code{data} and \code{diversity_data}
}
\description{
Generate Expertise Data Subsets
}
\details{
Note that this function needs to be run on \code{ManyEcoEvo} after the following functions have been run (See examples):
\itemize{
\item \code{prepare_response_variables()}
\item \code{generate_exclusion_subsets()}
\item \code{generate_rating_subsets()}
}

\code{generate_rating_subsets()} only creates expertise subsets based on the full dataset where \code{exclusion_set == "complete"} and \code{publishable_subset == "All"}.
}
\examples{
library(ManyEcoEvo)
library(tidyverse)
library(targets)
ManyEcoEvo \%>\%
prepare_response_variables(estimate_type = "Zr") |>
generate_exclusion_subsets(estimate_type = "Zr") |>
generate_rating_subsets() |>
generate_expertise_subsets(ManyEcoEvo:::expert_subset)
}
