% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rating_subsets.R
\name{generate_rating_subsets}
\alias{generate_rating_subsets}
\title{Generate subsets of ManyEcoEvo Data based on Peer Review Ratings}
\usage{
generate_rating_subsets(data)
}
\arguments{
\item{data}{a ManyEcoEvo dataframe containing formatted raw \code{data}, formatted \code{diversity_data}, the \code{estimate_type}, and \code{dataset}}
}
\value{
A ManyEcoEvo dataframe with added column \code{exclusion_set} with new subsets of \code{data} and \code{diversity_data}
}
\description{
Generates two subsets of data based on for both complete and partial exclusion datasets for both \code{yi} and \code{Zr} estimates.
}
\details{
To be executed on a ManyEcoEvo dataframe after the \code{compute_MA_inputs()} function. This function generates two subsets of data based on the peer review ratings for both complete and partial exclusion datasets for both \code{yi} and \code{Zr} estimates. The subsets are generated based on the \code{PublishableAsIs} column in the \code{data} list-column. The subsets are named \code{data_flawed} and \code{data_flawed_major} and are created by filtering out the data points with the \code{PublishableAsIs} values of \code{"deeply flawed and unpublishable"} and \code{"publishable with major revision"} respectively. The \code{diversity_data} list-column is also updated to reflect the new subsets of data.

Note, This function expects that within the list-column \code{data}, there is a list-column \code{review_data} containing the columns \code{ReviewerId}, \code{RateAnalysis}, \code{PublishableAsIs}, and \code{PriorBelief}.
}
\seealso{
Other Multi-dataset Wrapper Functions:
\code{\link[=apply_VZ_exclusions]{apply_VZ_exclusions()}},
\code{\link[=compute_MA_inputs]{compute_MA_inputs()}},
\code{\link[=generate_exclusion_subsets]{generate_exclusion_subsets()}},
\code{\link[=generate_outlier_subsets]{generate_outlier_subsets()}},
\code{\link[=generate_yi_subsets]{generate_yi_subsets()}},
\code{\link[=make_viz]{make_viz()}},
\code{\link[=meta_analyse_datasets]{meta_analyse_datasets()}},
\code{\link[=prepare_ManyEcoEvo]{prepare_ManyEcoEvo()}},
\code{\link[=prepare_ManyEcoEvo_yi]{prepare_ManyEcoEvo_yi()}},
\code{\link[=prepare_response_variables]{prepare_response_variables()}},
\code{\link[=prepare_response_variables_yi]{prepare_response_variables_yi()}},
\code{\link[=summarise_analysis_types]{summarise_analysis_types()}},
\code{\link[=summarise_conclusions]{summarise_conclusions()}},
\code{\link[=summarise_model_composition]{summarise_model_composition()}},
\code{\link[=summarise_reviews]{summarise_reviews()}},
\code{\link[=summarise_sorensen_index]{summarise_sorensen_index()}},
\code{\link[=summarise_variable_counts]{summarise_variable_counts()}}

Other targets-pipeline functions:
\code{\link[=prepare_ManyEcoEvo]{prepare_ManyEcoEvo()}},
\code{\link[=prepare_ManyEcoEvo_yi]{prepare_ManyEcoEvo_yi()}},
\code{\link[=prepare_response_variables_yi]{prepare_response_variables_yi()}},
\code{\link[=prepare_review_data]{prepare_review_data()}}
}
\concept{Multi-dataset Wrapper Functions}
\concept{targets-pipeline functions}
