% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_yi_subsets.R
\name{generate_yi_subsets}
\alias{generate_yi_subsets}
\title{Generate subsets of out-of-sample predictions data by \code{estimate_type} for multiple analysis datasets.}
\usage{
generate_yi_subsets(yi_analysis)
}
\arguments{
\item{yi_analysis}{A \code{dataframe} containing multiple datasets of out-of-sample predictions and corresponding diversity data.}
}
\value{
A \code{datraframe} with the character columns \code{dataset}, \code{estimate_type} and list-cols \code{data} and \code{diversity_data}
}
\description{
Generate subsets of out-of-sample predictions data by \code{estimate_type} for multiple analysis datasets.
}
\details{
The object \code{yi_analysis} should contain the character column \code{dataset}, and list-columns \code{data} and \code{diversity_data}. In essence, both \code{data} and \code{diversity_data} are nested by \code{dataset}. \code{data} should be a list-col of \code{grouped_df}, which should also contain the list-col \code{back_transformed_data}.

If the column \code{back_transformed_data} contains \code{NA} values, that row will be removed from the list-column \code{data}.
If the column \code{estimate_type} is present in \code{yi_analysis}, it will be removed, and replaced with the \code{estimate_type} derived from the \code{scenario} and \code{SurveyID} columns in each \code{back_transformed_data} dataset.
}
\seealso{
Other Multi-dataset Wrapper Functions:
\code{\link[=apply_VZ_exclusions]{apply_VZ_exclusions()}},
\code{\link[=compute_MA_inputs]{compute_MA_inputs()}},
\code{\link[=generate_exclusion_subsets]{generate_exclusion_subsets()}},
\code{\link[=generate_outlier_subsets]{generate_outlier_subsets()}},
\code{\link[=generate_rating_subsets]{generate_rating_subsets()}},
\code{\link[=make_viz]{make_viz()}},
\code{\link[=meta_analyse_datasets]{meta_analyse_datasets()}},
\code{\link[=prepare_ManyEcoEvo]{prepare_ManyEcoEvo()}},
\code{\link[=prepare_ManyEcoEvo_yi]{prepare_ManyEcoEvo_yi()}},
\code{\link[=prepare_response_variables]{prepare_response_variables()}},
\code{\link[=prepare_response_variables_yi]{prepare_response_variables_yi()}},
\code{\link[=summarise_analysis_types]{summarise_analysis_types()}},
\code{\link[=summarise_conclusions]{summarise_conclusions()}},
\code{\link[=summarise_model_composition]{summarise_model_composition()}},
\code{\link[=summarise_reviews]{summarise_reviews()}},
\code{\link[=summarise_sorensen_index]{summarise_sorensen_index()}},
\code{\link[=summarise_variable_counts]{summarise_variable_counts()}}
}
\author{
Elliot Gould
}
\concept{Multi-dataset Wrapper Functions}
