% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diversity_data.R
\name{get_diversity_data}
\alias{get_diversity_data}
\title{Get Diversity Data}
\usage{
get_diversity_data(raw_data, dataset, variables = character())
}
\arguments{
\item{raw_data}{}

\item{dataset}{character string of either "eucalyptus" or "blue tit"}

\item{variables}{character vector of any length containing names of variables to get diversity data for}
}
\value{
A tibble of diversity data for either 'eucalyptus' or 'blue tit' analyses
}
\description{
Get Diversity Data
}
\details{
Called by \code{prepare_diversity_raw}
}
