% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MA_fit_stats.R
\name{get_MA_fit_stats}
\alias{get_MA_fit_stats}
\title{Extract meta-analytic statistics like \eqn{I^2}, etc.}
\usage{
get_MA_fit_stats(MA_mod)
}
\arguments{
\item{MA_mod}{a fitted model of class rma.mv}
}
\value{
A tidy tibble with \eqn{\sigma^2} and \eqn{I^2} estimates for \code{MA_mod}
}
\description{
Extract meta-analytic statistics like \eqn{I^2}, etc.
}
