% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_forest.R
\name{gg_forest}
\alias{gg_forest}
\title{Forestplot with ggplot2}
\usage{
gg_forest(meta_model, estimate_type, dataset = character(1L))
}
\arguments{
\item{meta_model}{Fitted meta-analysis model of class \code{rma}}

\item{estimate_type}{Whether the estimate is standardized effect size of standardised out of sample prediction. Must be one of "Zr", "y50", "y25", "y75".}

\item{dataset}{Character string equal to the name of the dataset. One of "blue tit" or "eucalyptus"}
}
\value{
An object of class \code{ggplot2}
}
\description{
Forestplot with ggplot2
}
