% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i2_ml.R
\name{i2_ml}
\alias{i2_ml}
\title{i2_ml}
\usage{
i2_ml(model, method = c("ratio", "matrix"), data, boot = NULL)
}
\arguments{
\item{model}{Model object of class \code{rma.mv} or \code{rma}. Currently only model objects using the \code{mods} argument work (e.g., \code{mod = ~1}).}

\item{method}{Method used to calculate I2. Two options exist: a ratio-based calculation proposed by Nakagawa & Santos (\code{"ratio"}), or Wolfgang Viechtbauer's matrix method (\code{"matrix"}).}

\item{data}{Data frame used to fit the model.}

\item{boot}{Number of simulations to run to produce 95 percent confidence intervals for I2. Default is \code{NULL}, where only the point estimate is provided.}
}
\value{
A data frame containing all the model results including mean effect size estimate, confidence, and prediction intervals
}
\description{
I2 (I-squared) for mulilevel meta-analytic models, based on Nakagawa & Santos (2012). Under multilevel models, we can have multiple I2 (see also Senior et al. 2016). Alternatively, the method proposed by Wolfgang Viechtbauer can also be used.
}
\author{
Shinichi Nakagawa - s.nakagawa@unsw.edu.au

Daniel Noble - daniel.noble@anu.edu.au

Elliot Gould - elliot.gould@unimelb.edu.au
}
