% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_transform.R
\name{log_transform}
\alias{log_transform}
\title{log transform response-scale yi estimates}
\usage{
log_transform(estimate, std.error, sim)
}
\arguments{
\item{estimate}{Point estimates on the original response scale}

\item{std.error}{Standard error of the point estimates on the original response scale}

\item{sim}{Number of simulations to run when generating distribution to sample from}
}
\value{
A dataframe containing the \code{mean_log}, \code{std.error_log}, \code{lower} and \code{upper} 95\% CI on the log scale
}
\description{
log transform response-scale yi estimates
}
\examples{
log_transform(1.77, 1.01, 1000)
}
