% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_transform_yi.R
\name{log_transform_yi}
\alias{log_transform_yi}
\title{Log-transform a data-frame of back-transformed out-of-sample estimates}
\usage{
log_transform_yi(back_transformed_data, sim = 10000L)
}
\arguments{
\item{back_transformed_data}{description}

\item{sim}{Number of simulations to run when generating distribution to sample from}
}
\value{
A tibble of standardised-out-of-sample predictions on the Z-scale, with columns \code{Z}, \code{VZ}, \code{lower} and \code{upper} and the original columns fro \code{back_transformed_data} that were not used / updated in the transformation.
}
\description{
Log-transform a data-frame of back-transformed out-of-sample estimates
}
\seealso{
Equivalent to \code{\link[=pred_to_Z]{pred_to_Z()}} in terms of workflow data hierarchy.
}
