% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_outlier_subsets.R
\name{map_match_formulae}
\alias{map_match_formulae}
\title{Map matching formulae to data}
\usage{
map_match_formulae(data, variable_name, formulae, col_name = "outcome_colname")
}
\arguments{
\item{data}{A tibble}

\item{variable_name}{character string, the name of the variable to match}

\item{formulae}{A named list of formulae}
}
\value{
A tibble with a new column \code{outcome_colname} containing \code{rhs} of the the matched formula, or \code{NA} if no match is found
}
\description{
Map matching formulae to data
}
\keyword{internal}
