% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Negative Value Matching}
\usage{
lhs \%in\% rhs
}
\arguments{
\item{lhs}{vector or NULL: the values to be matched. \code{\link[base]{Long vectors}} are supported.}

\item{rhs}{vector or NULL: the values to be matched. \code{\link[base]{Long vectors}}  are supported.}
}
\value{
A logical vector indicating which value of \code{lhs} are \emph{not} matched in \code{rhs}
}
\description{
See \code{base::\href{`\%in\%`}{}} for details. \verb{\%nin\%} is a binary operator,
returning a logical vector indicating if there is a negative match or not.
}
\examples{
"A" \%nin\% LETTERS[1:10]
"A" \%in\% LETTERS[1:10]
}
\seealso{
Other Utility functions:
\code{\link[=capwords]{capwords()}},
\code{\link[=named_group_split]{named_group_split()}},
\code{\link[=rm_inf_na]{rm_inf_na()}},
\code{\link[=subset_fns_Zr]{subset_fns_Zr()}},
\code{\link[=subset_fns_yi]{subset_fns_yi()}}
}
\concept{Utility functions}
\keyword{internal}
