% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_response_variables.R
\name{pmap_wrap}
\alias{pmap_wrap}
\title{Wrapper function to standardise response variables}
\usage{
pmap_wrap(..., fns, env = caller_env())
}
\description{
Wrapper function to standardise response variables
}
\seealso{
This internal helper function is called by \code{\link[=prepare_response_variables]{prepare_response_variables()}}
}
\keyword{internal}
