% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sorensen_glm.R
\name{poss_fit_sorensen_glm}
\alias{poss_fit_sorensen_glm}
\title{Possibly \code{\link[=fit_sorensen_glm]{fit_sorensen_glm()}}}
\usage{
poss_fit_sorensen_glm(...)
}
\description{
A version of \code{\link[=fit_sorensen_glm]{fit_sorensen_glm()}} that returns \code{NA} if an error is encountered
}
\keyword{internal}
