% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_descriptive_statistics.R
\name{prepare_df_for_summarising}
\alias{prepare_df_for_summarising}
\title{Prepare data for summarising descriptive statistics}
\usage{
prepare_df_for_summarising(data)
}
\arguments{
\item{data}{A tibble containing many-analyst data to be summarised.}
}
\value{
A tibble containing the data prepared for summarising.
}
\description{
Calculates the number of fixed variables, the number of random variables,
the sample size, the number of interactions, the number of linear models,
the number of generalised models, the number of fixed effects, and the number
of random effects for each analysis in the dataset.
Also codes whether each analysis employs a linear model, a generalised model,
or a Bayesian model.
}
\examples{
ManyEcoEvo::ManyEcoEvo \%>\%
  select(data) \%>\%
  unnest(everything()) \%>\%
  prepare_df_for_summarising()
}
