% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_diversity_raw.R
\name{prepare_diversity_raw}
\alias{prepare_diversity_raw}
\title{Prepare diversity index data}
\usage{
prepare_diversity_raw(master_data_raw, master_raw_metadata)
}
\arguments{
\item{master_data_raw}{Raw dataset of both eucalyptus and blue tit data}

\item{master_raw_metadata}{Metadata describing the variable names in \code{master_data_raw}}
}
\value{
A tibble of raw diversity index data
}
\description{
Prepares data required for calculating Sorensen diversity indices across all analyses. \(Non-targets function\)
}
