% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_response_variables.R
\name{prepare_response_variables}
\alias{prepare_response_variables}
\title{Prepare response variable data for nested ManyEcoEvo dataset}
\usage{
prepare_response_variables(
  data,
  estimate_type = character(1L),
  param_table = NULL,
  dataset_standardise = NULL,
  dataset_log_transform = NULL,
  ...
)
}
\arguments{
\item{data}{Complete ManyEcoEvo dataset containing nested datasets for each different analysis and exclusion set dataset}

\item{estimate_type}{A character string of length 1, equal to either "Zr", "yi", "y25", "y50", "y75", indicating what type of estimates are being prepared.}

\item{param_table}{A table of parameters \(mean, sd\) for \emph{most} response variables used by analysts. This tibble is pulled from the named object exported by \verb{ManyEcoEvo::}. but can be overwritten with the users's own \code{param_table} dataset.}

\item{dataset_standardise}{A character string of length 1, equal to the name of the dataset to standardise the response variables to. If \code{NULL} (default), no datasets are standardised.}

\item{dataset_log_transform}{A character string of length 1, equal to the name of the dataset to log-transform the response variables to. If \code{NULL} (default), no datasets are log-transformed.}
}
\value{
A tibble of nested list-columns
}
\description{
Prepare response variable data for nested ManyEcoEvo dataset
}
\details{
Operates on nested list-columns of dataframes, where each dataframe contains the response variable data for a single analysis. The function standardises the response variable data for each analysis, and returns the modified dataset to the \code{data} list-column.

Note that if \code{data} does not have an \code{estimate_type} column, it will be added with the value of \code{estimate_type}. This is because some transformation functions require the \code{estimate_type} column to be present in the dataset.
}
\examples{
ManyEcoEvo \%>\% prepare_response_variables(estimate_type = "Zr")
}
\seealso{
Other Multi-dataset Wrapper Functions:
\code{\link[=apply_VZ_exclusions]{apply_VZ_exclusions()}},
\code{\link[=compute_MA_inputs]{compute_MA_inputs()}},
\code{\link[=generate_exclusion_subsets]{generate_exclusion_subsets()}},
\code{\link[=generate_outlier_subsets]{generate_outlier_subsets()}},
\code{\link[=generate_rating_subsets]{generate_rating_subsets()}},
\code{\link[=generate_yi_subsets]{generate_yi_subsets()}},
\code{\link[=make_viz]{make_viz()}},
\code{\link[=meta_analyse_datasets]{meta_analyse_datasets()}},
\code{\link[=prepare_ManyEcoEvo]{prepare_ManyEcoEvo()}},
\code{\link[=prepare_ManyEcoEvo_yi]{prepare_ManyEcoEvo_yi()}},
\code{\link[=prepare_response_variables_yi]{prepare_response_variables_yi()}},
\code{\link[=summarise_analysis_types]{summarise_analysis_types()}},
\code{\link[=summarise_conclusions]{summarise_conclusions()}},
\code{\link[=summarise_model_composition]{summarise_model_composition()}},
\code{\link[=summarise_reviews]{summarise_reviews()}},
\code{\link[=summarise_sorensen_index]{summarise_sorensen_index()}},
\code{\link[=summarise_variable_counts]{summarise_variable_counts()}}
}
\concept{Multi-dataset Wrapper Functions}
\concept{targets-pipeline functions.}
