% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_descriptive_statistics.R
\name{prepare_sorenson_summary_data}
\alias{prepare_sorenson_summary_data}
\title{Prepare data for summarising Sorensen diversity indices}
\usage{
prepare_sorenson_summary_data(
  data,
  data_subset_name = "all",
  id_subsets = list(),
  subset_names = character(0L),
  filter_expressions = NULL
)
}
\arguments{
\item{data}{A ManyAnalyst style tibble containing the data to be analysed.}

\item{data_subset_name}{A character vector of length 1, the name of the subset of \code{data}.}

\item{id_subsets}{A list of tibbles containing the \code{id_col} for each subset of \code{data}.}

\item{subset_names}{A character vector equal to the length of \code{id_subsets}; the name of data subsets in \code{id_subsets}.}

\item{filter_expressions}{A list of expressions to filter the data by.}
}
\value{
A tibble containing subsets of Sorensen diversity indices \code{data}.
}
\description{
Prepares the data for summarising Sorensen diversity indices across an entire study by unnesting the
diversity indices and joining them to the prepared data in preparation for
summarising subsets of data with \code{summarise_study()}.
}
\examples{
id_subsets <- list(ManyEcoEvo:::effect_ids, ManyEcoEvo:::prediction_ids)
subset_names <- c("effects", "predictions")
filter_vars <- rlang::exprs(
  exclusion_set == "complete",
  estimate_type == "Zr",
  publishable_subset == "All",
  expertise_subset == "All",
  collinearity_subset == "All"
)
prepare_sorenson_summary_data(ManyEcoEvo::ManyEcoEvo_results,
  "all",
  id_subsets,
  subset_names,
  filter_expressions = filter_vars
)
}
