% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_prediction_files.R
\name{preprocess_prediction_files}
\alias{preprocess_prediction_files}
\title{Preprocess Prediction Files}
\usage{
preprocess_prediction_files(predictions_validation, all_surveys)
}
\arguments{
\item{predictions_validation}{a dataframe containing the outputs of pointblank validation for multiple out of sample prediction files}

\item{all_surveys}{A tibble of analyst-data for all datasets}
}
\value{
pointblank check columns are nested into the list-column \code{checks}, \code{submission_id} is computed from \code{question}, \code{new_names} for replacement column names are given in \code{new_names}, \code{filepath} is appended with "inst/extdata/analyst_data/S2"
}
\description{
Preprocess Prediction Files
}
