% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_updated_prediction_files.R
\name{preprocess_updated_prediction_files}
\alias{preprocess_updated_prediction_files}
\title{Performs QA on re-submitted out-of-sample prediction files}
\usage{
preprocess_updated_prediction_files(df = data.frame())
}
\arguments{
\item{df}{A dataframe containing information about new prediction files}
}
\value{
A tibble with list-columns containing the outputs of point-blank checks
}
\description{
Performs QA on re-submitted out-of-sample prediction files
}
