% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_submission_data.R
\name{read_submission_data}
\alias{read_submission_data}
\title{Read out-of-sample-prediction analyst submission data}
\usage{
read_submission_data(filepath)
}
\value{
A tibble of out of sample prediction data
}
\description{
Read out-of-sample-prediction analyst submission data
}
