% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_predictions.R
\name{rename_prediction_cols}
\alias{rename_prediction_cols}
\title{Rename Prediction Columns}
\usage{
rename_prediction_cols(.data, key_var, .old_data)
}
\arguments{
\item{.data}{A tibble of out of sample predictions on the response variable scale}

\item{key_var}{A tibble of the key variables used in the conversion}

\item{.old_data}{A tibble of the original data used in the conversion}
}
\value{
A tibble of out of sample predictions on the response variable scale with the correct column names
}
\description{
Renames the prediction columns in the output of convert_predictions
}
\seealso{
Other Analysis-level functions: 
\code{\link{Z_VZ_preds}()},
\code{\link{back_transform_response_vars_yi}()},
\code{\link{box_cox_transform}()},
\code{\link{convert_predictions}()},
\code{\link{est_to_zr}()},
\code{\link{folded_params}()},
\code{\link{pmap_wrap}()},
\code{\link{variance_box_cox}()}

Other Back-transformation: 
\code{\link{back}},
\code{\link{back_transform_response_vars_yi}()},
\code{\link{conversion}()},
\code{\link{conversion_2}()},
\code{\link{convert_predictions}()}
}
\concept{Analysis-level functions}
\concept{Back-transformation}
