% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model_checks.R
\name{run_model_checks}
\alias{run_model_checks}
\title{Perform model checking on series of fitted models for different datasets, exclusion sets and estimate types}
\usage{
run_model_checks(data, model_name = character(0L))
}
\arguments{
\item{data}{A series of fitted models stored in list-columns for different \code{model_name}s, \code{exclusion_set}s, \code{dataset}s and \code{estimate_type}s}

\item{model_name}{character string of model name to be extracted from the nested data frame. String must be contained in \code{data$model_name}}
}
\value{
A tibble whose columns contain the outcomes of different model checks
}
\description{
Checks for convergence, singularity and heteroscedasticy
}
