% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_outlier_subsets.R
\name{slice_conditionally}
\alias{slice_conditionally}
\title{Slice Conditionally}
\usage{
slice_conditionally(data, n_min, n_max, outcome_variable)
}
\arguments{
\item{data}{A tibble}

\item{n_min}{integer, the number of bottom outliers to remove}

\item{n_max}{integer, the number of top outliers to remove}

\item{outcome_variable}{character string, the name of the outcome variable}
}
\description{
Slice Conditionally
}
\keyword{internal}
