% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_yi_subsets.R
\name{split_yi_subsets}
\alias{split_yi_subsets}
\title{Split a dataset of out-of-sample predictions by \code{estimate_type}}
\usage{
split_yi_subsets(.data)
}
\arguments{
\item{.data}{A dataset containing out-of-sample predictions}
}
\value{
A tibble of out-of-sample predictions subset by \code{estimate_type} and stored in the list-column \code{data}.
}
\description{
Reorganises the data from nesting based on individual analysis submissions, to nesting based on the type of estimate.
}
\details{
This function is used to split a dataset of out-of-sample predictions by \code{estimate_type}, where out-of-sample predictions are stored in a list-column called \code{back_transformed_data}, with one data frame of data per analysis submission.

The \code{estimate_type} is derived from either the \code{scenario} and \code{SurveyID} columns in each \code{back_transformed_data} dataset.
The \code{estimate_type} is then used to nest the data by \code{estimate_type} in the list-column \code{data}.

Removes unnecessary data in \code{.data}: \code{augmented_data} and \code{checks}.
}
\seealso{
This function is called in the wrapper function \code{\link[=generate_yi_subsets]{generate_yi_subsets()}} to split out-of-sample predictions by \code{estimate_type} across multiple ManyAnalyst datasets.
}
\author{
Elliot Gould
}
