% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_response.R
\name{standardise_response}
\alias{standardise_response}
\title{Standardise Response Variable}
\usage{
standardise_response(
  dat,
  estimate_type = character(1L),
  param_table = NULL,
  dataset = character(1L)
)
}
\arguments{
\item{estimate_type}{The type of estimate to be standardised. Character vector of length 1, whose value may be "Zr", "yi", "y25", "y50", "y75".}

\item{param_table}{A table of estimated 'population' parameters for each variable in the analysis datasets.}

\item{dataset}{One of either "blue tit" or "eucalyptus"}

\item{data}{}
}
\value{
A tibble of analyst data with standardised values contained in a list-column called 'back_transformed_data'
}
\description{
Standardise Response Variable
}
\details{
Note that for $y_i$ or out of sample predictions that are standardised, if param_table is \code{NA} or \code{NULL} for a given variable, then the response variable will not be standardised, and NA will be returned for that entry in \code{back_transformed_data}.
}
\seealso{
Other analyst-data: 
\code{\link{back_transform_response_vars_yi}()}
}
\concept{analyst-data}
