% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_fns_yi}
\alias{subset_fns_yi}
\title{Subsetting Functions for out-of-sample predictions meta-analysis}
\usage{
subset_fns_yi()
}
\value{
A named list of \code{lambda} functions
}
\description{
Generates a list of functions that are used to subset the
processed ManyEcoEvo dataset (e.g.\verb{data(ManyEcoEvo_results}) containing
out-of-sample predictions \eqn{y_i}.
}
\details{
The subset functions are used to filter out predictions that are not
to be included in the meta-analysis.
}
\seealso{
Other Utility functions:
\code{\link{\%nin\%}},
\code{\link[=capwords]{capwords()}},
\code{\link[=named_group_split]{named_group_split()}},
\code{\link[=rm_inf_na]{rm_inf_na()}},
\code{\link[=subset_fns_Zr]{subset_fns_Zr()}}
}
\concept{Utility functions}
