% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_fns_Zr}
\alias{subset_fns_Zr}
\title{Subsetting Functions for effect-size meta-analysis}
\usage{
subset_fns_Zr()
}
\value{
A named list of \code{lambda} functions
}
\description{
Generates a list of functions that are used to subset the processed \link{ManyEcoEvo} dataset
}
\seealso{
Other Utility functions:
\code{\link{\%nin\%}},
\code{\link[=capwords]{capwords()}},
\code{\link[=named_group_split]{named_group_split()}},
\code{\link[=rm_inf_na]{rm_inf_na()}},
\code{\link[=subset_fns_yi]{subset_fns_yi()}}
}
\concept{Utility functions}
