% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_review_data.R
\name{summarise_analyses_by_reviewer}
\alias{summarise_analyses_by_reviewer}
\title{Summarise analyses reviewed by reviewer}
\usage{
summarise_analyses_by_reviewer(review_data)
}
\arguments{
\item{review_data}{A dataframe containing the variables \code{dataset}, \code{response_id}, and \code{ReviewerId}.}
}
\value{
A dataframe summarising the \code{mean}, \code{sd}, \code{min}, and \code{max} of the number of reviews provided by each unique reviewer across \emph{all} analyses from all \code{dataset}s in \code{review_data}.
}
\description{
Summarise analyses reviewed by reviewer
}
\examples{
ManyEcoEvo \%>\%
ungroup \%>\%
select(data, -dataset) \%>\%
unnest(data) \%>\%
select(ends_with("_id"), id_col, dataset, review_data) \%>\%
unnest(review_data) \%>\%
summarise_analyses_by_reviewer()
}
