% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_conclusions.R
\name{summarise_conclusions_data}
\alias{summarise_conclusions_data}
\title{Count qualitative conclusions across all analyses for each dataset}
\usage{
summarise_conclusions_data(data)
}
\arguments{
\item{data}{A dataframe containing the columns \code{split_id}, \code{analysis_id}, \code{dataset}, \code{Conclusion}}
}
\value{
A dataframe with counts \code{n} for each unique value of \code{Conclusion} for each \code{dataset}
}
\description{
Count qualitative conclusions across all analyses for each dataset
}
\examples{
ManyEcoEvo$data[[1]] \%>\%
filter(Conclusion != "CHECK") \%>\%
summarise_conclusions_data()
}
