% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_sorensen_index.R
\name{summarise_sorensen_index_data}
\alias{summarise_sorensen_index_data}
\title{Summarise Sorensen's Mean Index Estimates for a dataframe}
\usage{
summarise_sorensen_index_data(data)
}
\arguments{
\item{data}{A dataframe containing \code{mean_diversity_index} for the Sorensen's index estimates for each analysis \code{id_col}, for each \code{dataset}.}
}
\value{
A dataframe with the \code{mean}, \code{sd}, \code{min}, \code{max} mean Sorensen's index values for each \code{dataset}.
}
\description{
Summarises Sorensen's index estimates for a single dataframe of estimates
}
\examples{
ManyEcoEvo_results \%>\%
filter(exclusion_set == "complete",
      publishable_subset == "All",
      expertise_subset == "All") \%>\%
 ungroup \%>\%
 select(dataset, diversity_indices) \%>\%
 unnest(diversity_indices) \%>\% 
 summarise_sorensen_index_data()
}
