% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_descriptive_statistics.R
\name{summarise_study}
\alias{summarise_study}
\title{Summarise ManyAnalyst study data}
\usage{
summarise_study(
  ManyEcoEvo,
  ManyEcoEvo_results,
  id_subsets,
  subset_names,
  filter_vars = NULL
)
}
\arguments{
\item{ManyEcoEvo}{A ManyAnalyst style tibble containing the data to be analysed.}

\item{ManyEcoEvo_results}{A ManyAnalyst results style tibble containing the results of the data to be analysed.}

\item{id_subsets}{A list of tibbles containing the \code{id_col} for each subset of \code{data}.}

\item{subset_names}{A character vector equal to the length of \code{id_subsets}; the name of data subsets in \code{id_subsets}.}

\item{filter_vars}{A list of expressions to filter \code{ManyEcoEvo_results} data}
}
\value{
A tibble containing summary statistics for each subset of data.
}
\description{
Summarises the ManyAnalyst study data by calculating summary statistics for each subset of data.
}
\examples{
id_subsets <- list(ManyEcoEvo:::effect_ids, ManyEcoEvo:::prediction_ids)
subset_names <- c("effects", "predictions")
filter_vars <- rlang::exprs(exclusion_set == "complete",
                            estimate_type == "Zr",
                            publishable_subset == "All",
                            expertise_subset == "All",
                            collinearity_subset == "All")
summarise_study(ManyEcoEvo::ManyEcoEvo, ManyEcoEvo::ManyEcoEvo_results, id_subsets, subset_names, filter_vars = filter_vars)
}
