% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_variable_counts.R
\name{summarise_variable_counts}
\alias{summarise_variable_counts}
\title{Summarise variable usage across analyses}
\usage{
summarise_variable_counts(
  ManyEcoEvo,
  ManyEcoEvo_results,
  ManyEcoEvo_yi_results,
  output = "count"
)
}
\arguments{
\item{ManyEcoEvo}{A tibble of \code{ManyEcoEvo}}

\item{ManyEcoEvo_results}{A tibble of \code{ManyEcoEvo_results}}

\item{ManyEcoEvo_yi_results}{A tibble of \code{ManyEcoEvo_yi_results}}

\item{output}{A length 1 character vector equal to \code{"count"} or \code{"aggregate"}}
}
\value{
A dataframe of count values \code{n} or summary statistic values \code{n_mean},\code{n_sd},\code{n_min},\code{n_max} of counts depending on the the value supplied to the \code{output} argument.
}
\description{
Using \code{\link[=count_analyses_variables_used]{count_analyses_variables_used()}} calculates the counts of analyses in which each variable is used for a given \code{dataset}. Further summarisation of those counts ( \code{mean}, \code{sd}, \code{min} and \code{max}) is optional, see details.
}
\details{
To calculate count values, supply \code{"count"} to argument \code{output}. To calculate summary statistics of counts, supply \code{"aggregate"} to argument \code{output} instead.
}
\examples{
summarise_variable_counts(ManyEcoEvo, ManyEcoEvo_results, ManyEcoEvo_yi_results, "count")

summarise_variable_counts(ManyEcoEvo, ManyEcoEvo_results, ManyEcoEvo_yi_results, "aggregate")
}
\seealso{
Other Multi-dataset Wrapper Functions:
\code{\link[=apply_VZ_exclusions]{apply_VZ_exclusions()}},
\code{\link[=compute_MA_inputs]{compute_MA_inputs()}},
\code{\link[=generate_exclusion_subsets]{generate_exclusion_subsets()}},
\code{\link[=generate_outlier_subsets]{generate_outlier_subsets()}},
\code{\link[=generate_rating_subsets]{generate_rating_subsets()}},
\code{\link[=generate_yi_subsets]{generate_yi_subsets()}},
\code{\link[=make_viz]{make_viz()}},
\code{\link[=meta_analyse_datasets]{meta_analyse_datasets()}},
\code{\link[=prepare_ManyEcoEvo]{prepare_ManyEcoEvo()}},
\code{\link[=prepare_ManyEcoEvo_yi]{prepare_ManyEcoEvo_yi()}},
\code{\link[=prepare_response_variables]{prepare_response_variables()}},
\code{\link[=prepare_response_variables_yi]{prepare_response_variables_yi()}},
\code{\link[=summarise_analysis_types]{summarise_analysis_types()}},
\code{\link[=summarise_conclusions]{summarise_conclusions()}},
\code{\link[=summarise_model_composition]{summarise_model_composition()}},
\code{\link[=summarise_reviews]{summarise_reviews()}},
\code{\link[=summarise_sorensen_index]{summarise_sorensen_index()}}
}
\author{
Hannah S. Fraser

Elliot Gould
}
\concept{Multi-dataset Wrapper Functions}
