% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_predictions.R
\name{validate}
\alias{validate}
\alias{validate_predictions_df_blue_tit}
\alias{validate_predictions_df_euc}
\alias{validate_predictions}
\title{Validating analyst-submitted predictions}
\usage{
validate_predictions_df_blue_tit(input, type = "filepath")

validate_predictions_df_euc(input, type = "filepath")

validate_predictions(data_set, input, type = "filepath")
}
\arguments{
\item{input}{Either a filepath or a dataframe, corresponding to the \code{type} argument specification}

\item{type}{character string of length 1, equal to either "filepath" or "df". Defaults to "filepath"}

\item{data_set}{the dataset being analysed, either "blue tit" or "eucalyptus"}
}
\value{
A pointblank agent, class \code{ptblank_agent}, yet to be interrogated
}
\description{
Validates the structure of the analyst-submitted predictions, ensuring that the required columns are present and in the correct format.
}
\section{Functions}{
\itemize{
\item \code{validate_predictions_df_blue_tit()}: Validate Blue tit predictions data

\item \code{validate_predictions_df_euc()}: Validate \emph{Eucalyptus} predictions data

\item \code{validate_predictions()}: Wrapper-function for \code{\link[=validate_predictions_df_euc]{validate_predictions_df_euc()}} and \code{\link[=validate_predictions_df_blue_tit]{validate_predictions_df_blue_tit()}}

}}
\seealso{
\code{\link[pointblank:create_agent]{pointblank::create_agent()}}
}
\concept{Out-of-Sample Prediction Validation}
