% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_predictions.R
\name{validate_predictions}
\alias{validate_predictions}
\title{Validate predictions conditioned on dataset}
\usage{
validate_predictions(data_set, input, type = "filepath")
}
\arguments{
\item{data_set}{the dataset being analysed, either "blue tit" or "eucalyptus"}

\item{input}{a tibble of out of sample predictions that has been augmented, i.e. checked and updated}

\item{type}{character string of length 1, equal to either "filepath" or "df". Defaults to "filepath"}
}
\value{
An object of class "pointblank_agent", yet to be interrogated.
}
\description{
Validate predictions conditioned on dataset
}
