% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_predictions.R
\name{validate_predictions_df_blue_tit}
\alias{validate_predictions_df_blue_tit}
\title{Validate out of sample predictions for blue tit dataset with pointblank package}
\usage{
validate_predictions_df_blue_tit(input, type = "filepath")
}
\arguments{
\item{input}{Either a filepath or a dataframe, corresponding to the \code{type} argument specification}

\item{type}{character string of length 1, equal to either "filepath" or "df". Defaults to "filepath"}
}
\value{
A pointblank agent, class "ptblank_agent", yet to be interrogated
}
\description{
Validate out of sample predictions for blue tit dataset with pointblank package
}
