% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Z_VZ_preds.R
\name{Z_VZ_preds}
\alias{Z_VZ_preds}
\title{Standardize out-of-sample predictions}
\usage{
Z_VZ_preds(yi, yi_se, mu_p, sd_p)
}
\arguments{
\item{yi}{point-estimate prediction, on response scale}

\item{yi_se}{standard error of \code{yi}}

\item{mu_p}{Population mean for variable estimated by \code{yi}}

\item{sd_p}{Population standard deviation for variable estimated by \code{yi}}
}
\value{
A tibble with columns \code{Z} and \code{VZ}
}
\description{
Standardize out-of-sample predictions
}
\seealso{
Other Analysis-level functions:
\code{\link[=apply_sorensen_calc]{apply_sorensen_calc()}},
\code{\link[=back_transform_response_vars_yi]{back_transform_response_vars_yi()}},
\code{\link[=box_cox_transform]{box_cox_transform()}},
\code{\link[=convert_predictions]{convert_predictions()}},
\code{\link[=est_to_zr]{est_to_zr()}},
\code{\link[=folded_params]{folded_params()}},
\code{\link[=pred_to_Z]{pred_to_Z()}},
\code{\link[=variance_box_cox]{variance_box_cox()}}
}
\concept{Analysis-level functions}
