% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_sorensen_calc.R
\name{apply_sorensen_calc}
\alias{apply_sorensen_calc}
\title{Applies the sorensen diversity index calculation to variable diversity dataset}
\usage{
apply_sorensen_calc(.data)
}
\arguments{
\item{.data}{A variable diversity dataset with columns for each unique variable used in all analyses. Values are \code{NA} if the analysis did not use that variable, and take the value of the variable as a character string if the analysis did use that variable. Each unique analysis is a row.}
}
\value{
A tibble containing the variables \code{id_col}, \code{mean_diversity_index} and \code{num_variables}
}
\description{
Applies the sorensen diversity index calculation to variable diversity dataset
}
