% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_VZ_exclusions.R
\name{apply_VZ_exclusions}
\alias{apply_VZ_exclusions}
\title{Apply VZ exclusion to a data-frame containing list-columns of yi subsets}
\usage{
apply_VZ_exclusions(df = data.frame(), VZ_colname, VZ_cutoff)
}
\arguments{
\item{df}{A dataframe of yi data subsets generated by \code{\link[=generate_yi_subsets]{generate_yi_subsets()}} or \code{\link[=split_yi_subsets]{split_yi_subsets()}}.}

\item{VZ_colname}{Either A character vector of length 1, the name of the column in the dataframes stored in \code{df}'s list-column \code{data} that contains the VZ values. Or else a named list of character values, where the names are the dataset names and the values are the \code{VZ_colname}s  for each \code{dataset} in \code{df}.}

\item{VZ_cutoff}{A numeric vector of length 1, values equal to or greater than this value of VZ will be filtered out of the dataframes stored in \code{df}'s list-column \code{data}, else a named list of numeric values, where the names are the dataset names and the values are the \code{VZ_cutoff}s for each \code{dataset} in \code{df}.}
}
\value{
A dataframe of yi subsets, whose extreme values of VZ have been removed.
}
\description{
Apply VZ exclusion to a data-frame containing list-columns of yi subsets
}
\details{
\code{df} must contain the columns \code{"data"}, \code{"diversity_data"} and \code{"dataset"}.
If only one value of \code{VZ_colname} and \code{VZ_cutoff} is supplied, it will be recycled to match the number of datasets in \code{df}.

If a named list is supplied for \code{VZ_colname} and \code{VZ_cutoff}, the names must match the dataset names in \code{df}.
}
\examples{
data(ManyEcoEvo_yi)
ManyEcoEvo_yi \%>\%
  prepare_response_variables(
    estimate_type = "yi",
    param_table =
      ManyEcoEvo:::analysis_data_param_tables,
    dataset_standardise = "blue tit",
    dataset_log_transform = "eucalyptus") \%>\%
  generate_yi_subsets() \%>\% 
  apply_VZ_exclusions(VZ_colname = 
                        list("eucalyptus" = "se_log", 
                             "blue tit" = "VZ"), 
                      VZ_cutoff = 3)
}
\seealso{
Applies \code{\link[=exclude_extreme_VZ]{exclude_extreme_VZ()}} to each dataframe in the list-column \code{data} of \code{df}.

Other Multi-dataset Wrapper Functions: 
\code{\link{compute_MA_inputs}()},
\code{\link{generate_exclusion_subsets}()},
\code{\link{generate_outlier_subsets}()},
\code{\link{generate_rating_subsets}()},
\code{\link{generate_yi_subsets}()},
\code{\link{make_viz}()},
\code{\link{meta_analyse_datasets}()},
\code{\link{prepare_ManyEcoEvo}()},
\code{\link{prepare_ManyEcoEvo_yi}()},
\code{\link{prepare_response_variables}()},
\code{\link{prepare_response_variables_yi}()},
\code{\link{summarise_analysis_types}()},
\code{\link{summarise_conclusions}()},
\code{\link{summarise_model_composition}()},
\code{\link{summarise_reviews}()},
\code{\link{summarise_sorensen_index}()},
\code{\link{summarise_variable_counts}()}
}
\concept{Multi-dataset Wrapper Functions}
