% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_VZ_exclusions.R
\name{apply_VZ_exclusions}
\alias{apply_VZ_exclusions}
\title{Apply VZ exclusion to a data-frame containing list-columns of yi subsets}
\usage{
apply_VZ_exclusions(df = data.frame(), VZ_cutoff = numeric(1L))
}
\arguments{
\item{df}{A dataframe of yi data subsets generated from \verb{generate_yi_subsets\\(\\)}.}

\item{VZ_cutoff}{A numeric vector of length 1, values equal to or greater than this value of VZ will be filtered out of the dataframes stored in \code{df}'s list-column \code{data}.}
}
\value{
A dataframe of yi subsets, whose extreme values of VZ have been removed.
}
\description{
Apply VZ exclusion to a data-frame containing list-columns of yi subsets
}
\seealso{
Other Multi-dataset Wrapper Functions: 
\code{\link{compute_MA_inputs}()},
\code{\link{generate_exclusion_subsets}()},
\code{\link{generate_outlier_subsets}()},
\code{\link{generate_rating_subsets}()},
\code{\link{generate_yi_subsets}()},
\code{\link{make_viz}()},
\code{\link{meta_analyse_datasets}()},
\code{\link{prepare_ManyEcoEvo}()},
\code{\link{prepare_ManyEcoEvo_yi}()},
\code{\link{prepare_response_variables}()},
\code{\link{prepare_response_variables_yi}()},
\code{\link{summarise_analysis_types}()},
\code{\link{summarise_conclusions}()},
\code{\link{summarise_model_composition}()},
\code{\link{summarise_reviews}()},
\code{\link{summarise_sorensen_index}()},
\code{\link{summarise_variable_counts}()}
}
\concept{Multi-dataset Wrapper Functions}
