% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/back_transformations.R
\name{back}
\alias{back}
\alias{log_back}
\alias{logit_back}
\alias{probit_back}
\alias{inverse_back}
\alias{square_back}
\alias{cube_back}
\alias{identity_back}
\alias{power_back}
\alias{divide_back}
\alias{square_root_back}
\title{Back-transform effect-sizes to response scale.}
\usage{
log_back(beta, se, sim)

logit_back(beta, se, sim)

probit_back(beta, se, sim)

inverse_back(beta, se, sim)

square_back(beta, se, sim)

cube_back(beta, se, sim)

identity_back(beta, se, sim)

power_back(beta, se, sim, n)

divide_back(beta, se, sim, n)

square_root_back(beta, se, sim)
}
\arguments{
\item{beta}{Analyst beta estimate}

\item{se}{Standard error of analyst's effect size estimate \eqn{\beta}
or out-of-sample prediction estimate \eqn{y_i}.}

\item{sim}{numeric vector of length 1. number of simulations.}

\item{n}{Denominator used by analyst to divide the response variable.}
}
\value{
data frame containing the mean estimate, its standard error, and quantiles.
}
\description{
Transforms effect-sizes and their standard errors to the response scale.
}
\details{
We assume analysts' estimates are normally distributed. Each function uses a normal distribution to simulate the a distribution of effect-sizes and their standard errors. Next this distribution is back-transformed to the desired response scale. The mean \code{m_est}, standard error \code{se_est}, and quantiles (\code{lower} and \code{upper}) of the back-transformed distribution are returned within a dataframe.
}
\section{Functions}{
\itemize{
\item \code{log_back()}: Back transform beta estimates for models with log-link

\item \code{logit_back()}: Back transform beta estimates for models with logit-link

\item \code{probit_back()}: Back transform beta estimates for models with probit-link

\item \code{inverse_back()}: Back transform beta estimates for models with \eqn{1/x} link

\item \code{square_back()}: Back transform beta estimates for models with \eqn{x^2}-link

\item \code{cube_back()}: Back transform beta estimates for models with \eqn{x^3}-link

\item \code{identity_back()}: Back transform beta estimates for models with identity-link

\item \code{power_back()}: Back transform beta estimates for models with power-link

\item \code{divide_back()}: Back transform beta estimates or out-of-sample predictions from models whose response variable has been divided by some number, \code{n}.

\item \code{square_root_back()}: Back transform beta estimates or out-of-sample predictions from models whose response variable has been transformed by the square root

}}
\seealso{
Other Back-transformation: 
\code{\link{assign_transformation_type}()},
\code{\link{back_transform_response_vars_yi}()},
\code{\link{clean_response_transformation}()},
\code{\link{conversion}()},
\code{\link{conversion_2}()},
\code{\link{convert_predictions}()},
\code{\link{rename_prediction_cols}()}

Other Back-transformation: 
\code{\link{assign_transformation_type}()},
\code{\link{back_transform_response_vars_yi}()},
\code{\link{clean_response_transformation}()},
\code{\link{conversion}()},
\code{\link{conversion_2}()},
\code{\link{convert_predictions}()},
\code{\link{rename_prediction_cols}()}

Other Back-transformation: 
\code{\link{assign_transformation_type}()},
\code{\link{back_transform_response_vars_yi}()},
\code{\link{clean_response_transformation}()},
\code{\link{conversion}()},
\code{\link{conversion_2}()},
\code{\link{convert_predictions}()},
\code{\link{rename_prediction_cols}()}

Other Back-transformation: 
\code{\link{assign_transformation_type}()},
\code{\link{back_transform_response_vars_yi}()},
\code{\link{clean_response_transformation}()},
\code{\link{conversion}()},
\code{\link{conversion_2}()},
\code{\link{convert_predictions}()},
\code{\link{rename_prediction_cols}()}

Other Back-transformation: 
\code{\link{assign_transformation_type}()},
\code{\link{back_transform_response_vars_yi}()},
\code{\link{clean_response_transformation}()},
\code{\link{conversion}()},
\code{\link{conversion_2}()},
\code{\link{convert_predictions}()},
\code{\link{rename_prediction_cols}()}

Other Back-transformation: 
\code{\link{assign_transformation_type}()},
\code{\link{back_transform_response_vars_yi}()},
\code{\link{clean_response_transformation}()},
\code{\link{conversion}()},
\code{\link{conversion_2}()},
\code{\link{convert_predictions}()},
\code{\link{rename_prediction_cols}()}

Other Back-transformation: 
\code{\link{assign_transformation_type}()},
\code{\link{back_transform_response_vars_yi}()},
\code{\link{clean_response_transformation}()},
\code{\link{conversion}()},
\code{\link{conversion_2}()},
\code{\link{convert_predictions}()},
\code{\link{rename_prediction_cols}()}

Other Back-transformation: 
\code{\link{assign_transformation_type}()},
\code{\link{back_transform_response_vars_yi}()},
\code{\link{clean_response_transformation}()},
\code{\link{conversion}()},
\code{\link{conversion_2}()},
\code{\link{convert_predictions}()},
\code{\link{rename_prediction_cols}()}

Other Back-transformation: 
\code{\link{assign_transformation_type}()},
\code{\link{back_transform_response_vars_yi}()},
\code{\link{clean_response_transformation}()},
\code{\link{conversion}()},
\code{\link{conversion_2}()},
\code{\link{convert_predictions}()},
\code{\link{rename_prediction_cols}()}

Other Back-transformation: 
\code{\link{assign_transformation_type}()},
\code{\link{back_transform_response_vars_yi}()},
\code{\link{clean_response_transformation}()},
\code{\link{conversion}()},
\code{\link{conversion_2}()},
\code{\link{convert_predictions}()},
\code{\link{rename_prediction_cols}()}

Other Back-transformation: 
\code{\link{assign_transformation_type}()},
\code{\link{back_transform_response_vars_yi}()},
\code{\link{clean_response_transformation}()},
\code{\link{conversion}()},
\code{\link{conversion_2}()},
\code{\link{convert_predictions}()},
\code{\link{rename_prediction_cols}()}
}
\concept{Back-transformation}
