% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_response_variables_yi.R
\name{back_transform_response_vars_yi}
\alias{back_transform_response_vars_yi}
\title{Back Transform Response Variables - yi}
\usage{
back_transform_response_vars_yi(
  dat,
  estimate_type = character(1L),
  dataset = character(1L)
)
}
\arguments{
\item{dat}{A dataframe of out of sample predictions analyst submission data}

\item{estimate_type}{The type of estimate to be standardised. Character vector of length 1, whose value may be  "yi", "y25", "y50", "y75".}

\item{dataset}{One of either "blue tit" or "eucalyptus"}
}
\value{
A tibble of analyst data with standardised values contained in a list-column called 'back_transformed_data'
}
\description{
Back Transform Response Variables - yi
}
\seealso{
Other Analysis-level functions: 
\code{\link{Z_VZ_preds}()},
\code{\link{apply_sorensen_calc}()},
\code{\link{box_cox_transform}()},
\code{\link{convert_predictions}()},
\code{\link{est_to_zr}()},
\code{\link{folded_params}()},
\code{\link{pmap_wrap}()},
\code{\link{rename_prediction_cols}()},
\code{\link{variance_box_cox}()}

Other Back-transformation: 
\code{\link{assign_transformation_type}()},
\code{\link{back}},
\code{\link{clean_response_transformation}()},
\code{\link{conversion}()},
\code{\link{conversion_2}()},
\code{\link{convert_predictions}()},
\code{\link{rename_prediction_cols}()}
}
\concept{Analysis-level functions}
\concept{Back-transformation}
