% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_response_variables_yi.R
\name{back_transform_response_vars_yi}
\alias{back_transform_response_vars_yi}
\title{Back-transform analyst \code{yi} estiamtes to the response scale}
\usage{
back_transform_response_vars_yi(data)
}
\arguments{
\item{data}{dataframe of out of sample predictions analyst submission data}
}
\value{
A tibble of analyst data with standardised values contained in a list-column called 'back_transformed_data'
}
\description{
Back-transform analyst \code{yi} estiamtes to the response scale
}
\details{
\code{data} must contain the columns: \code{id_col}, \code{augmented_data}, \code{transformation}, \code{response_transformation_status}, where \code{augmented_data} is a list-column of dataframes containing the analyst predictions, \code{transformation} is the transformation used in the model, and \code{response_transformation_status} is the status of the response transformation.
}
\seealso{
Other Analysis-level functions: 
\code{\link{Z_VZ_preds}()},
\code{\link{apply_sorensen_calc}()},
\code{\link{box_cox_transform}()},
\code{\link{convert_predictions}()},
\code{\link{est_to_zr}()},
\code{\link{folded_params}()},
\code{\link{pmap_wrap}()},
\code{\link{variance_box_cox}()}

Other Back-transformation: 
\code{\link{assign_transformation_type}()},
\code{\link{back}},
\code{\link{clean_response_transformation}()},
\code{\link{conversion}()},
\code{\link{conversion_2}()},
\code{\link{convert_predictions}()}
}
\concept{Analysis-level functions}
\concept{Back-transformation}
