% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_cox_transform.R
\name{box_cox_transform}
\alias{box_cox_transform}
\title{Box-cox transform absolute deviation from the meta-analytic mean scores}
\usage{
box_cox_transform(data, dataset)
}
\arguments{
\item{data}{Dataset for model fitting, must contain columns \code{"abs_deviation_score_estimate"} and standard error}

\item{dataset}{character string of either "blue tit" or "eucalyptus"}
}
\value{
data with additional columns of box-cox transformed deviation scores and variance
}
\description{
Box-cox transform absolute deviation from the meta-analytic mean scores
}
\examples{
# library(tidyverse)
# library(targets)
# source("R/functions.R")
# tar_load(meta_analysis_outputs)
# box_cox_transform(data = meta_analysis_outputs$data[[1]],
#                   dataset = meta_analysis_outputs$dataset[[1]])
}
