% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_deviation_score.R
\name{calculate_deviation_score}
\alias{calculate_deviation_score}
\title{Calculate deviation from meta-analytic mean}
\usage{
calculate_deviation_score(data, meta_analytic_model, outcome_colname)
}
\arguments{
\item{data}{Dataset containing the column \code{"Z"} \(standardised out of sample prediction piont estimates\) or standardised effect-sizes, \code{"Zr"}.}

\item{meta_analytic_model}{Fitted meta-analytic model of class \code{rma}}

\item{outcome_colname}{Column name in \code{data} containing outcomes used in meta-analysis}
}
\value{
\code{data} but with an additional column of deviation score estimates \code{"abs_deviation_score_estimate"}
}
\description{
Calculate the absolute deviation of the standardised effect size from the meta-analytic mean
}
