% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sorensen_diversity_index.R
\name{calculate_sorensen_diversity_index}
\alias{calculate_sorensen_diversity_index}
\title{Calculate mean Sorensen pair-wise dissimilarity values for a ManyAnalyst dataset}
\usage{
calculate_sorensen_diversity_index(.data, .id = character())
}
\arguments{
\item{.data}{A variable diversity dataset with columns for each unique variable used in all analyses. Values are \code{NA} if the analysis did not use that variable, and take the value of the variable as a character string if the analysis did use that variable. Each unique analysis is a row.}

\item{.id}{character string of the analysis identifier column}
}
\value{
A tibble containing the variables \code{id_col}, \code{mean_diversity_index} and \code{num_variables}; the total number of variables used in a given analysis.
}
\description{
Calculate mean Sorensen pair-wise dissimilarity values for a ManyAnalyst dataset
}
\details{
Sorensen pair-wise dissimilarity is calculated as the number of shared variables between two analyses divided by the total number of variables used in both analyses. After calculating the Sorensen dissimilarity for all pairwise comparisons, the function then computes the mean diversity index for each analysis as the mean of all pairwise comparisons for that analysis, generating a single value per analysis.
}
\seealso{
See \code{\link[betapart:beta.pair]{betapart::beta.pair()}} for details of the Sorensen dissimilarity calculation.
}
