% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capwords}
\alias{capwords}
\title{Capitalise Words}
\usage{
capwords(s, strict = FALSE)
}
\arguments{
\item{s}{character string containing words to be captalised}

\item{strict}{whether to capitalise all words or not}
}
\value{
a character string whose words are now capitalised
}
\description{
Capitalise Words
}
\examples{
capwords("bah, bah, black sheep")
}
