% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_response_transformation.R
\name{clean_response_transformation}
\alias{clean_response_transformation}
\title{Clean response transformation variable}
\usage{
clean_response_transformation(response_transformation)
}
\arguments{
\item{response_transformation}{A character vector with the}
}
\value{
A character vector of cleaned response transformation values equal to the required \code{transformation} values in \code{conversion()}
}
\description{
Cleans the response transformation variable to the names of the back-transformation functions
}
\concept{back-transformation functions}
