% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_MA_inputs.R
\name{compute_MA_inputs}
\alias{compute_MA_inputs}
\title{Compute meta-analysis inputs for a nested-dataframe containing different datasets/subsets of analyst data}
\usage{
compute_MA_inputs(ManyEcoEvo, estimate_type = NULL)
}
\arguments{
\item{ManyEcoEvo}{A dataframe grouped by the character columns \code{dataset}, \code{estimate_type}, \code{exclusion_set}. Each group corresponds to a subset of the full \code{dataset}, and has the subset analyst data  stored in \code{data}, with its corresponding subset \code{diversity_data}.}

\item{estimate_type}{}
}
\value{
A dataframe that includes the additional columns in \code{ManyEcoEvo}, but with added columns \code{diversity_indices} and \code{effects_analysis}.
}
\description{
Computes the sorensen diversity indices and joins it to the prepared data in preparation for meta-analysing all subsets of data with \code{meta_analyse_datasets()}.
}
\details{
The name of the subset is derived from the funs generated by \code{subset_fns_yi} and/or \code{subset_fns_Zr}.

Computes sorensen diversity indices \code{diversity_indices} for each subset of data returning them in the list-columns \code{diversity_indices} and joins them to the subset analyst \code{data}, returning it in the list-column \code{effects_analysis}.
}
\seealso{
Other Multi-dataset Wrapper Functions: 
\code{\link{apply_VZ_exclusions}()},
\code{\link{generate_exclusion_subsets}()},
\code{\link{generate_outlier_subsets}()},
\code{\link{generate_rating_subsets}()},
\code{\link{generate_yi_subsets}()},
\code{\link{make_viz}()},
\code{\link{meta_analyse_datasets}()},
\code{\link{prepare_ManyEcoEvo_yi}()},
\code{\link{prepare_ManyEcoEvo}()},
\code{\link{prepare_response_variables_yi}()},
\code{\link{prepare_response_variables}()},
\code{\link{summarise_analysis_types}()},
\code{\link{summarise_conclusions}()},
\code{\link{summarise_model_composition}()}
}
\concept{Multi-dataset Wrapper Functions}
