% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_predictions.R
\name{convert_predictions}
\alias{convert_predictions}
\title{Convert Predictions}
\usage{
convert_predictions(
  augmented_data,
  transformation_type,
  response_transformation,
  link_fun
)
}
\arguments{
\item{augmented_data}{A tibble of out of analysts's sample prediction data for 3 scenarios}
}
\value{
A tibble of out of sample predictions on the response variable scale of the response variable used by the analyst
}
\description{
Converts out of sample predictions on the link scale back to the response scale
}
\seealso{
Other analysis-values: 
\code{\link{est_to_zr}()}
}
\concept{analysis-values}
