% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_boxcox_ratings_cat.R
\name{fit_boxcox_ratings_cat}
\alias{fit_boxcox_ratings_cat}
\title{Fit model of boxcox deviation scores as function of continuous ratings}
\usage{
fit_boxcox_ratings_cat(.data, outcome, outcome_var, interceptless = FALSE)
}
\arguments{
\item{.data}{Data for model fitting}

\item{outcome}{outcome variable, unquoted.}

\item{outcome_var}{Variance of the \code{outcome} variable}

\item{interceptless}{A logical relating to whether the model should be interceptless or not. Defaults to \code{FALSE}.}
}
\value{
An object of class lmer.
}
\description{
Fit an lmer model of the box-cox transformed deviation from the meta-analytic mean scores as a function of continuous peer-review ratings
}
