% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_boxcox_ratings_cont.R
\name{fit_boxcox_ratings_cont}
\alias{fit_boxcox_ratings_cont}
\title{Fit model of boxcox deviation scores as function of continuous ratings}
\usage{
fit_boxcox_ratings_cont(
  .data,
  outcome,
  outcome_var,
  ...,
  env = rlang::caller_env()
)
}
\arguments{
\item{.data}{Data for model fitting}

\item{outcome}{outcome variable, unquoted.}

\item{outcome_var}{Variance of the \code{outcome} variable}
}
\value{
An object of class lmer.
}
\description{
Fit an lmer model of the box-cox transformed deviation from the meta-analytic mean scores as a function of continuous peer-review ratings
}
\examples{
# Example Usage:
# Note, outcome = the name using NSE of the response variable, otucome_var = the
# variance associated with that variable.
# library(tidyverse);library(targets);library(metafor);library(tidymodels);library(multilevelmod)
# tar_load(meta_analysis_outputs)
# meta_analysis_outputs$data[[1]] \%>\%
#   fit_boxcox_ratings_cont(.,
#                                   box_cox_abs_deviation_score_estimate,
#                                   VZr )
}
