% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_MA_mv.R
\name{fit_MA_mv}
\alias{fit_MA_mv}
\title{Fit multivariate meta-analysis of effect-sizes}
\usage{
fit_MA_mv(
  effects_analysis = data.frame(),
  Z_colname,
  VZr_colname,
  estimate_type = character(1L)
)
}
\arguments{
\item{effects_analysis}{A dataframe containing containing effect-sizes  and their variances.}

\item{Z_colname}{Name of the variable containing standardised effect-sizes.}

\item{estimate_type}{character string, one of "Zr", "yi", "y25", "y50", "y75".}

\item{VZ_colname}{Name of the variable containing the variance of standardised effect-sizes.}
}
\value{
A fitted model of class \code{rma.mv} and \code{rma}.
}
\description{
Applies \code{\link[=fit_metafor_mv]{fit_metafor_mv()}} to fit a multivariate meta-analysis model using the \code{metafor} package. Nested random effects are included for \code{TeamIdentifier} and \code{TeamIdentifier/study_id }.
}
\examples{
ManyEcoEvo_results$effects_analysis[[1]] \%>\% 
fit_MA_mv(beta_estimate, beta_SE, "Zr")
}
