% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_metafor_mv.R
\name{fit_metafor_mv}
\alias{fit_metafor_mv}
\title{Fit Multivariate Metaregression using metafoR}
\usage{
fit_metafor_mv(estimate, variance, estimate_type = character(1L), data)
}
\arguments{
\item{estimate}{Numeric vector}

\item{variance}{Numeric vector}

\item{estimate_type}{Character vector of either "Zr", "y50", "y25", "y75".}

\item{data}{Dataframe containing estimates and variances with case id column \code{study_id}.}
}
\value{
Object of class \code{rma.mv}
}
\description{
Fit Multivariate Metaregression using metafoR
}
\seealso{
Other Model fitting and meta-analysis: 
\code{\link{fit_MA_mv}()},
\code{\link{fit_boxcox_ratings_cat}()},
\code{\link{fit_boxcox_ratings_cont}()},
\code{\link{fit_boxcox_ratings_ord}()},
\code{\link{fit_metafor_mv_reduced}()},
\code{\link{fit_metafor_uni}()},
\code{\link{fit_multivar_MA}()},
\code{\link{fit_sorensen_glm}()},
\code{\link{fit_uni_mixed_effects}()},
\code{\link{poss_fit_boxcox_ratings_cat}()},
\code{\link{poss_fit_boxcox_ratings_cont}()},
\code{\link{poss_fit_boxcox_ratings_ord}()},
\code{\link{poss_fit_metafor_mv}()},
\code{\link{poss_fit_uni_mixed_effects}()}
}
\concept{Model fitting and meta-analysis}
