% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_metafor_mv.R
\name{fit_metafor_mv}
\alias{fit_metafor_mv}
\title{Fit Multivariate Metaregression using metafoR}
\usage{
fit_metafor_mv(estimate, variance, estimate_type = character(1L), data)
}
\arguments{
\item{estimate}{Numeric vector}

\item{variance}{Numeric vector}

\item{estimate_type}{Character vector of either "Zr", "y50", "y25", "y75".}

\item{data}{Dataframe containing estimates and variances with case id column \code{study_id}.}
}
\value{
Object of class \code{rma.mv}
}
\description{
Fit Multivariate Metaregression using metafoR
}
\examples{
# TODO -- is this the best way of setting up this fun?? (i.e. to take numeric vectors)?
# Example Usage:
# library(tidyverse);library(targets);library(metafor) # NOT RUN, TODO: remove after create pkg
# source("R/functions.R") #NOT RUN, TODO: remove after create pkg
tar_read(round_2_survey_meta_analysis) \%>\%
  filter(dataset == "eucalyptus") \%>\%
  filter(
    !is.na(Zr),
    !is.na(VZr),
    !is.infinite(Zr),
    !is.infinite(VZr)
  ) \%>\%
  fit_metafor_mv(estimate = .$Zr, variance = .$VZr, estimate_type = "Zr", data = .)
}
