% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sorensen_glm.R
\name{fit_sorensen_glm}
\alias{fit_sorensen_glm}
\title{Fit univariate glm of deviation scores on sorensen diversity index}
\usage{
fit_sorensen_glm(data)
}
\arguments{
\item{data}{A dataframe containing box-cox transformed absolute deviation from the meta-analytic mean scores}
}
\value{
An fitted model object of class \code{glm} and \code{parsnip}
}
\description{
Fit univariate glm of deviation scores on sorensen diversity index
}
\examples{
# library(tidyverse);library(targets);library(metafor);library(tidymodels)
# tar_load(meta_analysis_outputs)
# fit_sorensen_glm(meta_analysis_results$data[[1]])
}
