% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sorensen_glm.R
\name{fit_sorensen_glm}
\alias{fit_sorensen_glm}
\title{Fit univariate glm of deviation scores on sorensen diversity index}
\usage{
fit_sorensen_glm(data)
}
\arguments{
\item{data}{A dataframe containing box-cox transformed absolute deviation from the meta-analytic mean scores}
}
\value{
An fitted model object of class \code{glm} and \code{parsnip}
}
\description{
Fit univariate glm of deviation scores on sorensen diversity index
}
\examples{
# library(tidyverse);library(targets);library(metafor);library(tidymodels)
# tar_load(meta_analysis_outputs)
# fit_sorensen_glm(meta_analysis_results$data[[1]])
}
\seealso{
Other Model fitting and meta-analysis: 
\code{\link{fit_MA_mv}()},
\code{\link{fit_boxcox_ratings_cat}()},
\code{\link{fit_boxcox_ratings_cont}()},
\code{\link{fit_boxcox_ratings_ord}()},
\code{\link{fit_metafor_mv}()},
\code{\link{fit_metafor_mv_reduced}()},
\code{\link{fit_metafor_uni}()},
\code{\link{fit_multivar_MA}()},
\code{\link{fit_uni_mixed_effects}()},
\code{\link{poss_fit_boxcox_ratings_cat}()},
\code{\link{poss_fit_boxcox_ratings_cont}()},
\code{\link{poss_fit_boxcox_ratings_ord}()},
\code{\link{poss_fit_metafor_mv}()},
\code{\link{poss_fit_uni_mixed_effects}()}
}
\concept{Model fitting and meta-analysis}
