% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_cox_transform.R
\name{folded_params}
\alias{folded_params}
\title{Calculate the folded parameters for the Box-Cox transformation}
\usage{
folded_params(abs_dev_score, VZr)
}
\arguments{
\item{abs_dev_score}{The absolute deviation score}

\item{VZr}{The variance of the standardised effect size}
}
\value{
A list containing the mean and variance of the folded parameters
}
\description{
Calculate the folded parameters for the Box-Cox transformation
}
\seealso{
Other Box-Cox transformation: 
\code{\link{box_cox_transform}()},
\code{\link{variance_box_cox}()}

Other Analysis-level functions: 
\code{\link{Z_VZ_preds}()},
\code{\link{apply_sorensen_calc}()},
\code{\link{back_transform_response_vars_yi}()},
\code{\link{box_cox_transform}()},
\code{\link{convert_predictions}()},
\code{\link{est_to_zr}()},
\code{\link{pmap_wrap}()},
\code{\link{rename_prediction_cols}()},
\code{\link{variance_box_cox}()}
}
\concept{Analysis-level functions}
\concept{Box-Cox transformation}
