% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_exclusion_subsets.R
\name{generate_exclusion_subsets}
\alias{generate_exclusion_subsets}
\title{Generate subsets of analyst data based on different exclusion criteria}
\usage{
generate_exclusion_subsets(data, estimate_type = NULL)
}
\arguments{
\item{data}{A dataframe containing at minimum the character column \code{dataset} and list-columns \code{data} and \code{diversity_data}.}

\item{estimate_type}{character vector, one of \code{"Zr", "yi", "y25", "y50", "y75", NULL}.}
}
\value{
A \code{dataframe} grouped by \code{dataset} and \code{exclusion_set} that contains subsets of \code{data} and \code{diversity_data} based on exclusion criteria functions defined in \code{\link[=subset_fns_Zr]{subset_fns_Zr()}} and \code{\link[=subset_funs_yi]{subset_funs_yi()}}, and the \code{estimate_type} column.
}
\description{
Generates subsets of data with different combinations of outliers removed for different \code{exclusion_set}s.
}
\details{
If \code{estimate_type} is \code{NULL}, the column \code{estimate_type} must be in \code{data}.

This function uses the functions \code{\link[=subset_fns_Zr]{subset_fns_Zr()}} and []subset_fns_yi() to create named lists with elements containing \verb{purrr::}-style lambda functions, whose element name is the name of the function.

If \code{NULL} is provided to the argument \code{estimate_type}, then the column \code{estimate_type} must exist in \code{data}, as this column will be used to filter the exclusion criteria functions.

The value of column \code{exclusion_set} in the returned object will be the name of the exclusion criteria function, with the prefix "subset_" removed, derived from either \code{\link[=subset_fns_Zr]{subset_fns_Zr()}} or \code{\link[=subset_funs_yi]{subset_funs_yi()}}, depending on the \code{estimate_type}.

\strong{Note}: This function should be exectued after \code{\link[=prepare_response_variables]{prepare_response_variables()}} and \code{\link[=generate_yi_subsets]{generate_yi_subsets()}}.
}
\seealso{
Other Multi-dataset Wrapper Functions: 
\code{\link{apply_VZ_exclusions}()},
\code{\link{compute_MA_inputs}()},
\code{\link{generate_outlier_subsets}()},
\code{\link{generate_rating_subsets}()},
\code{\link{generate_yi_subsets}()},
\code{\link{make_viz}()},
\code{\link{meta_analyse_datasets}()},
\code{\link{prepare_ManyEcoEvo}()},
\code{\link{prepare_ManyEcoEvo_yi}()},
\code{\link{prepare_response_variables}()},
\code{\link{prepare_response_variables_yi}()},
\code{\link{summarise_analysis_types}()},
\code{\link{summarise_conclusions}()},
\code{\link{summarise_model_composition}()},
\code{\link{summarise_reviews}()},
\code{\link{summarise_sorensen_index}()},
\code{\link{summarise_variable_counts}()}
}
\concept{Multi-dataset Wrapper Functions}
