% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_exclusion_subsets.R
\name{generate_exclusion_subsets}
\alias{generate_exclusion_subsets}
\title{Generate subsets of analyst data based on different exclusion criteria}
\usage{
generate_exclusion_subsets(ManyEcoEvo, estimate_type = NULL)
}
\arguments{
\item{ManyEcoEvo}{A dataframe}

\item{estimate_type}{character vector, one of \code{"Zr", "yi", "y25", "y50", "y75", NULL}.}
}
\value{
A \code{dataframe} grouped by dataset and exclusion_set that contains subsets of \code{data} and \code{diversity_data} based on exclusion criteria functions defined in \code{subset_fns_Zr} and \code{subset_funs_yi},
}
\description{
Generate subsets of analyst data based on different exclusion criteria
}
\details{
If \code{estimate_type} is \code{NULL}, the column \code{estimate_type} must be in \code{ManyEcoEvo}.
}
\note{
This function uses the functions\code{subset_fns_Zr}and \code{subset_fns_yi} to create named lists with elements containing \code{rlang_lambda_function}s, whose element name is the name of the \code{rlang_lambda_function}

To be run after \code{generate_exclusion_subsets()}
}
\seealso{
Other Multi-dataset Wrapper Functions: 
\code{\link{apply_VZ_exclusions}()},
\code{\link{compute_MA_inputs}()},
\code{\link{generate_outlier_subsets}()},
\code{\link{generate_rating_subsets}()},
\code{\link{generate_yi_subsets}()},
\code{\link{make_viz}()},
\code{\link{meta_analyse_datasets}()},
\code{\link{prepare_ManyEcoEvo_yi}()},
\code{\link{prepare_ManyEcoEvo}()},
\code{\link{prepare_response_variables_yi}()},
\code{\link{prepare_response_variables}()},
\code{\link{summarise_analysis_types}()},
\code{\link{summarise_conclusions}()},
\code{\link{summarise_model_composition}()},
\code{\link{summarise_variable_counts}()}
}
\concept{Multi-dataset Wrapper Functions}
