% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rating_subsets.R
\name{generate_rating_subsets}
\alias{generate_rating_subsets}
\title{Generate subsets of ManyEcoEvo Data based on Peer Review Ratings}
\usage{
generate_rating_subsets(ManyEcoEvo)
}
\arguments{
\item{ManyEcoEvo}{a ManyEcoEvo dataframe containing formatted raw \code{data}, formatted \code{diversity_data}, the \code{estimate_type}, and \code{dataset}}
}
\value{
A ManyEcoEvo dataframe with added column \code{exclusion_set} with new subsets of \code{data} and \code{diversity_data}
}
\description{
Generates two subsets of data based on for both complete and partial exclusion datasets for both \code{yi} and \code{Zr} estimates.
}
\seealso{
Other Multi-dataset Wrapper Functions: 
\code{\link{apply_VZ_exclusions}()},
\code{\link{compute_MA_inputs}()},
\code{\link{generate_exclusion_subsets}()},
\code{\link{generate_outlier_subsets}()},
\code{\link{generate_yi_subsets}()},
\code{\link{make_viz}()},
\code{\link{meta_analyse_datasets}()},
\code{\link{prepare_ManyEcoEvo_yi}()},
\code{\link{prepare_ManyEcoEvo}()},
\code{\link{prepare_response_variables_yi}()},
\code{\link{prepare_response_variables}()},
\code{\link{summarise_analysis_types}()},
\code{\link{summarise_conclusions}()},
\code{\link{summarise_model_composition}()},
\code{\link{summarise_variable_counts}()}

Other targets-pipeline functions: 
\code{\link{generate_outlier_subsets}()},
\code{\link{make_viz}()},
\code{\link{prepare_ManyEcoEvo_yi}()},
\code{\link{prepare_ManyEcoEvo}()},
\code{\link{prepare_response_variables_yi}()},
\code{\link{prepare_response_variables}()},
\code{\link{prepare_review_data}()}
}
\concept{Multi-dataset Wrapper Functions}
\concept{targets-pipeline functions}
