% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forest_plot_data.R
\name{get_forest_plot_data}
\alias{get_forest_plot_data}
\title{Get Forest Plot Data from a Metafor Model}
\usage{
get_forest_plot_data(model)
}
\arguments{
\item{model}{A metafor model object of class \code{rma.mv} or \code{rma.uni}}
}
\value{
A tibble containing the data required to plot a forest plot
}
\description{
Get Forest Plot Data from a Metafor Model
}
\examples{
get_forest_plot_data(model)
}
\seealso{
Other Plotting functions: 
\code{\link{plot_cont_rating_effects}()},
\code{\link{plot_effects_diversity}()},
\code{\link{plot_forest}()},
\code{\link{plot_model_means_box_cox_cat}()},
\code{\link{plot_model_means_orchard}()}
}
\concept{Plotting functions}
