% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_viz.R
\name{make_viz}
\alias{make_viz}
\title{Make visualisations wrapper function}
\usage{
make_viz(data)
}
\arguments{
\item{data}{a nested dataframe with processed and standardised data stored in list-column \code{data}, grouped by variables \code{exclusion_set}, \code{dataset}, \code{estimate_type}}
}
\value{
a nested dataframe grouped by variables \code{exclusion_set}, \code{dataset}, \code{estimate_type},
}
\description{
Computes model summaries, tidy model summaries, model fit stats, funnel plots and forest plots for a dataframe of multiple fitted models
}
\seealso{
Other targets-pipeline functions: 
\code{\link{generate_outlier_subsets}()},
\code{\link{generate_rating_subsets}()},
\code{\link{prepare_ManyEcoEvo}()},
\code{\link{prepare_ManyEcoEvo_yi}()},
\code{\link{prepare_response_variables_yi}()},
\code{\link{prepare_review_data}()}

Other Multi-dataset Wrapper Functions: 
\code{\link{apply_VZ_exclusions}()},
\code{\link{compute_MA_inputs}()},
\code{\link{generate_exclusion_subsets}()},
\code{\link{generate_outlier_subsets}()},
\code{\link{generate_rating_subsets}()},
\code{\link{generate_yi_subsets}()},
\code{\link{meta_analyse_datasets}()},
\code{\link{prepare_ManyEcoEvo}()},
\code{\link{prepare_ManyEcoEvo_yi}()},
\code{\link{prepare_response_variables}()},
\code{\link{prepare_response_variables_yi}()},
\code{\link{summarise_analysis_types}()},
\code{\link{summarise_conclusions}()},
\code{\link{summarise_model_composition}()},
\code{\link{summarise_reviews}()},
\code{\link{summarise_sorensen_index}()},
\code{\link{summarise_variable_counts}()}
}
\concept{Multi-dataset Wrapper Functions}
\concept{targets-pipeline functions}
