% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManyEcoEvo-package.R
\docType{package}
\name{ManyEcoEvo-package}
\alias{ManyEcoEvo}
\alias{ManyEcoEvo-package}
\title{ManyEcoEvo: Meta-analyse data from 'Many-Analysts' style studies}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The ManyEcoEvo package provides a set of functions for aiding data cleaning, preparation and conducting analysis for multi-analyst style studies. It also contains datasets from the ManyEcoEvo project for reproducing the ManyEcoEvo analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/egouldo/ManyEcoEvo}
  \item \url{https://egouldo.github.io/ManyEcoEvo/}
  \item Report bugs at \url{https://github.com/egouldo/ManyEcoEvo/issues}
}

}
\author{
\strong{Maintainer}: Elliot Gould \email{elliot.gould@unimelb.edu.au} (\href{https://orcid.org/0000-0002-6585-538X}{ORCID})

Authors:
\itemize{
  \item Hannah S. Fraser (\href{https://orcid.org/0000-0003-2443-4463}{ORCID})
  \item Shinichi Nakagawa (\href{https://orcid.org/0000-0002-7765-5182}{ORCID})
  \item Timothy H. Parker (\href{https://orcid.org/0000-0003-2995-5284}{ORCID})
}

}
\keyword{internal}
