% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Negative Value Matching}
\usage{
lhs \%in\% rhs
}
\arguments{
\item{lhs}{vector or NULL: the values to be matched. \href{http://127.0.0.1:59782/help/library/base/help/Long\%20vectors}{Long vectors} are supported.}

\item{rhs}{vector or NULL: the values to be matched. \href{http://127.0.0.1:59782/help/library/base/help/Long\%20vectors}{Long vectors} are supported.}
}
\value{
A logical vector indicating which value of \code{lhs} are \emph{not} matched in \code{rhs}
}
\description{
See \code{base::\href{`\%in\%`}{}} for details. \verb{\%nin\%} is a binary operator,
returning a logical vector indicating if there is a negative match or not.
}
\examples{
"A" \%nin\% LETTERS[1:10]
"A" \%in\% LETTERS[1:10]
}
\keyword{internal}
