% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effects_diversity.R
\name{plot_effects_diversity}
\alias{plot_effects_diversity}
\title{Marginal Effects Plot of Diversity Index Model}
\usage{
plot_effects_diversity(mod, df, back_transform = FALSE)
}
\arguments{
\item{mod}{A fitted model of class 'lm'}

\item{df}{A dataframe with the columns, lambda, abs_deviation_score_estimate, box_cox_abs_deviation_score_estimate, mean_diversity_index}

\item{back_transform}{logical, to back-transform box-cox transformed to absolute deviation scores or not}
}
\value{
A ggplot with the original data,  predicted values and confidence values from the fitted model
}
\description{
Marginal Effects Plot of Diversity Index Model
}
\examples{
#targets::tar_load(ManyEcoEvo_results) #TODO change this to package data
#library(tidyverse)
#plot_effects_diversity(mod = ManyEcoEvo_results$sorensen_glm[[5]], 
#df = ManyEcoEvo_results$effects_analysis[[5]] \%>\% #dat
#select(mean_diversity_index, study_id, 
#          starts_with("box_cox"), 
#                   starts_with("abs_dev"),
#                            lambda), 
#                            back_transform = TRUE) 
}
