% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forest_plot_data.R
\name{plot_forest}
\alias{plot_forest}
\title{Plot a Forest Plot}
\usage{
plot_forest(data, intercept = TRUE, MA_mean = TRUE)
}
\arguments{
\item{data}{A tibble containing the data required to plot a forest plot}

\item{intercept}{Logical. Should a horizontal line be added at 0?}

\item{MA_mean}{Logical. Should a dashed line be added at the meta-analytic mean?}
}
\value{
A ggplot object
}
\description{
Plot a forest plot using the data from \code{get_forest_plot_data}
}
\examples{
data(ManyEcoEvo_results)
model <- ManyEcoEvo_results \%>\% pluck("MA_mod", 1) 
plot_data <- get_forest_plot_data(model)
plot_forest(plot_data)
plot_forest(plot_data, intercept = FALSE)
plot_forest(plot_data, MA_mean = FALSE)
plot_forest(plot_data, intercept = FALSE, MA_mean = FALSE)
}
\seealso{
Other Plotting functions: 
\code{\link{get_forest_plot_data}()},
\code{\link{plot_cont_rating_effects}()},
\code{\link{plot_effects_diversity}()},
\code{\link{plot_model_means_box_cox_cat}()},
\code{\link{plot_model_means_orchard}()}
}
\concept{Plotting functions}
