% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_means_box_cox_cat.R
\name{plot_model_means_box_cox_cat}
\alias{plot_model_means_box_cox_cat}
\title{plot_model_means_box_cox_cat}
\usage{
plot_model_means_box_cox_cat(
  dat,
  variable,
  predictor_means,
  new_order,
  title,
  lambda,
  back_transform = FALSE
)
}
\arguments{
\item{dat}{Data for plotting}

\item{variable}{Categorical predictor variable to plot}

\item{predictor_means}{A tibble containing the means and confidence intervals of the predictor variable}

\item{new_order}{A character vector of the new order of the levels of the categorical predictor}

\item{title}{A character vector of the plot title}

\item{lambda}{A length 1 numeric vector of the lambda value used in the box-cox transformation}

\item{back_transform}{A logical indicating whether to back-transform the box-cox transformed data}
}
\value{
A ggplot object
}
\description{
Plot model means for box-cox transformed deviation scores as a function of categorical ratings
}
