% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_boxcox_ratings_ord.R
\name{poss_fit_boxcox_ratings_ord}
\alias{poss_fit_boxcox_ratings_ord}
\title{Possibly \code{\link[=fit_box_cox_ratings]{fit_box_cox_ratings()}}}
\usage{
poss_fit_boxcox_ratings_ord(...)
}
\description{
A wrapper for \code{\link[=fit_box_cox_ratings]{fit_box_cox_ratings()}} that returns \code{NA} if an error is thrown
}
\seealso{
Other Model fitting and meta-analysis: 
\code{\link{fit_MA_mv}()},
\code{\link{fit_boxcox_ratings_cat}()},
\code{\link{fit_boxcox_ratings_cont}()},
\code{\link{fit_boxcox_ratings_ord}()},
\code{\link{fit_metafor_mv}()},
\code{\link{fit_metafor_mv_reduced}()},
\code{\link{fit_metafor_uni}()},
\code{\link{fit_multivar_MA}()},
\code{\link{fit_sorensen_glm}()},
\code{\link{fit_uni_mixed_effects}()},
\code{\link{poss_fit_boxcox_ratings_cat}()},
\code{\link{poss_fit_boxcox_ratings_cont}()},
\code{\link{poss_fit_metafor_mv}()},
\code{\link{poss_fit_uni_mixed_effects}()}
}
\concept{Model fitting and meta-analysis}
\keyword{internal}
