% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_metafor_mv.R
\name{poss_fit_metafor_mv}
\alias{poss_fit_metafor_mv}
\title{Possibly \code{\link[=fit_metafor_mv]{fit_metafor_mv()}}}
\usage{
poss_fit_metafor_mv(...)
}
\description{
Wrapper for \code{\link[=fit_metafor_mv]{fit_metafor_mv()}} that returns \code{NA} if an error is thrown
}
\seealso{
\code{\link[=fit_metafor_mv]{fit_metafor_mv()}}

Other Model fitting and meta-analysis: 
\code{\link{fit_MA_mv}()},
\code{\link{fit_boxcox_ratings_cat}()},
\code{\link{fit_boxcox_ratings_cont}()},
\code{\link{fit_boxcox_ratings_ord}()},
\code{\link{fit_metafor_mv}()},
\code{\link{fit_metafor_mv_reduced}()},
\code{\link{fit_metafor_uni}()},
\code{\link{fit_multivar_MA}()},
\code{\link{fit_sorensen_glm}()},
\code{\link{fit_uni_mixed_effects}()},
\code{\link{poss_fit_boxcox_ratings_cat}()},
\code{\link{poss_fit_boxcox_ratings_cont}()},
\code{\link{poss_fit_boxcox_ratings_ord}()},
\code{\link{poss_fit_uni_mixed_effects}()}
}
\concept{Model fitting and meta-analysis}
\keyword{internal}
