% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_to_Z.R
\name{pred_to_Z}
\alias{pred_to_Z}
\title{Standardize Out-Of-Sample Predictions}
\usage{
pred_to_Z(back_transformed_data, params, dataset)
}
\arguments{
\item{back_transformed_data}{a dataframe or tibble with the columns "estimate" and "se.fit", containing yi and SE\(yi\) values respectively}

\item{response_variable_name}{a character vector}
}
\description{
Standardizes out-of-sample predictions by computing the Fisher's Z transformed Correlation Coefficient from analysts' out-of-sample prediction estimates and corresponding standard error.
}
\note{
\code{pred_to_Z} expects estimates to be on the response scale, not the link scale.
}
