% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ManyEcoEvo.R
\name{prepare_ManyEcoEvo}
\alias{prepare_ManyEcoEvo}
\title{Prepare ManyEcoEvo raw dataset}
\usage{
prepare_ManyEcoEvo(master_data_raw, master_raw_metadata, all_review_data)
}
\arguments{
\item{master_data_raw}{Raw data for both eucalyptus and blue tit analyses}

\item{master_raw_metadata}{Metadata describing \code{master_data_raw}}

\item{all_review_data}{Review data for all \code{dataset}s contained in \code{master_data_raw}, generated by \code{prepare_review_data}.}
}
\value{
A tibble of ManyEcoEvo analyst response data for effect-sizes ($Z_r$).
}
\description{
Prepare ManyEcoEvo raw dataset
}
\seealso{
Other Multi-dataset Wrapper Functions: 
\code{\link{apply_VZ_exclusions}()},
\code{\link{compute_MA_inputs}()},
\code{\link{generate_exclusion_subsets}()},
\code{\link{generate_outlier_subsets}()},
\code{\link{generate_rating_subsets}()},
\code{\link{generate_yi_subsets}()},
\code{\link{make_viz}()},
\code{\link{meta_analyse_datasets}()},
\code{\link{prepare_ManyEcoEvo_yi}()},
\code{\link{prepare_response_variables_yi}()},
\code{\link{prepare_response_variables}()},
\code{\link{summarise_analysis_types}()},
\code{\link{summarise_conclusions}()},
\code{\link{summarise_model_composition}()},
\code{\link{summarise_variable_counts}()}

Other targets-pipeline functions: 
\code{\link{generate_outlier_subsets}()},
\code{\link{generate_rating_subsets}()},
\code{\link{make_viz}()},
\code{\link{prepare_ManyEcoEvo_yi}()},
\code{\link{prepare_response_variables_yi}()},
\code{\link{prepare_response_variables}()},
\code{\link{prepare_review_data}()}
}
\concept{Multi-dataset Wrapper Functions}
\concept{targets-pipeline functions}
