% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ManyEcoEvo_yi.R
\name{prepare_ManyEcoEvo_yi}
\alias{prepare_ManyEcoEvo_yi}
\title{Prepare ManyEcoEvo raw dataset for out-of-sample predictions}
\usage{
prepare_ManyEcoEvo_yi(
  master_data_raw,
  master_raw_metadata,
  all_prediction_data
)
}
\arguments{
\item{master_data_raw}{Raw data for both eucalyptus and blue tit analyses}

\item{master_raw_metadata}{Metadata describing \code{master_data_raw}}

\item{all_prediction_data}{Prediction data for all \code{dataset}s contained in a \code{tibble}}

\item{all_review_data}{Review data for all \code{dataset}s contained in \code{master_data_raw}, generated by \code{prepare_review_data}.}

\item{ManyEcoEvo}{ManyEcoEvo data generated by \code{prepare_ManyEcoEvo}}
}
\value{
A tibble of ManyEcoEvo analyst response data with out-of-sample predictions ($y_i$).
}
\description{
Prepare ManyEcoEvo raw dataset for out-of-sample predictions
}
\seealso{
Other targets-pipeline functions: 
\code{\link{generate_outlier_subsets}()},
\code{\link{generate_rating_subsets}()},
\code{\link{make_viz}()},
\code{\link{prepare_ManyEcoEvo}()},
\code{\link{prepare_response_variables_yi}()},
\code{\link{prepare_response_variables}()},
\code{\link{prepare_review_data}()}

Other Multi-dataset Wrapper Functions: 
\code{\link{apply_VZ_exclusions}()},
\code{\link{compute_MA_inputs}()},
\code{\link{generate_exclusion_subsets}()},
\code{\link{generate_outlier_subsets}()},
\code{\link{generate_rating_subsets}()},
\code{\link{generate_yi_subsets}()},
\code{\link{make_viz}()},
\code{\link{meta_analyse_datasets}()},
\code{\link{prepare_ManyEcoEvo}()},
\code{\link{prepare_response_variables_yi}()},
\code{\link{prepare_response_variables}()},
\code{\link{summarise_analysis_types}()},
\code{\link{summarise_conclusions}()},
\code{\link{summarise_model_composition}()}
}
\concept{Multi-dataset Wrapper Functions}
\concept{targets-pipeline functions}
