% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_review_data.R
\name{prepare_review_data}
\alias{prepare_review_data}
\title{Prepare peer-review data from Qualtrics}
\usage{
prepare_review_data(bt_reviews, euc_reviews)
}
\arguments{
\item{bt_reviews}{Blue tit review}

\item{euc_reviews}{}
}
\value{
a tibble of peer-review data for blue tit and eucalyptus analyses
}
\description{
Prepare peer-review data from Qualtrics
}
\seealso{
Other targets-pipeline functions: 
\code{\link{generate_outlier_subsets}()},
\code{\link{generate_rating_subsets}()},
\code{\link{make_viz}()},
\code{\link{prepare_ManyEcoEvo}()},
\code{\link{prepare_ManyEcoEvo_yi}()},
\code{\link{prepare_response_variables}()},
\code{\link{prepare_response_variables_yi}()}
}
\concept{targets-pipeline functions}
