% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rm_inf_na}
\alias{rm_inf_na}
\title{Removes infinite and NA values from a dataframe of standardised effects}
\usage{
rm_inf_na(effects_analysis, Z_colname, VZ_colname)
}
\arguments{
\item{effects_analysis}{}

\item{Z_colname}{unquoted or bare column name with the Z or Zr estimates}

\item{VZ_colname}{unquoted or bare column name containing the VZ or VZr estimates}
}
\value{
a dataframe without
}
\description{
Removes infinite and NA values from a dataframe of standardised effects
}
