% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rm_inf_na}
\alias{rm_inf_na}
\title{Removes infinite and NA values from a dataframe of standardised effects}
\usage{
rm_inf_na(effects_analysis, Z_colname, VZ_colname)
}
\arguments{
\item{effects_analysis}{A dataframe containing the standardised effects}

\item{Z_colname}{unquoted or bare column name with the \eqn{Z} or \eqn{Z_r} estimates}

\item{VZ_colname}{unquoted or bare column name containing the \eqn{VZ} or \eqn{\text{VZ}_r} estimates}
}
\value{
a dataframe without
}
\description{
Removes infinite and NA values from a dataframe of standardised effects
}
\examples{
data(ManyEcoEvo_results)
ManyEcoEvo_results \%>\% 
pluck("effects_analysis", 1) \%>\% 
rm_inf_na(beta_estimate, beta_SE)
}
