% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_fns_yi}
\alias{subset_fns_yi}
\title{Subsetting Functions for yi analysis}
\usage{
subset_fns_yi()
}
\value{
A named list of \code{lambda} functions
}
\description{
Generates a list of functions that are used to subset the processed ManyEcoEvo dataset containing out-of-sample predictions \(\code{y\_i}\)
}
