% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_fns_Zr}
\alias{subset_fns_Zr}
\title{Subsetting Functions for Zr analysis}
\usage{
subset_fns_Zr()
}
\value{
A named list of \code{lambda} functions
}
\description{
Generates a list of functions that are used to subset the processed ManyEcoEvo dataset
}
