% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_analysis_types.R
\name{summarise_analysis_types}
\alias{summarise_analysis_types}
\title{Summarise Analysis Types}
\usage{
summarise_analysis_types(ManyEcoEvo_results, ManyEcoEvo_yi_results, ManyEcoEvo)
}
\arguments{
\item{data}{}
}
\value{
A summarised tibble with the variables \code{subset}, \code{dataset}, \code{num_teams}, \code{total_analyses}, \code{sum_linear}, \code{sum_mixed}, \code{sum_Bayesian}.
}
\description{
Generates a summary of the number of analysis teams, total analyses, models with normal error distributions, mixed effects modls, and models developed using Bayesian statistical methods for a given analysis type.
}
\examples{
summarise_analysis_types(ManyEcoEvo_results, ManyEcoEvo_yi_results, ManyEcoEvo)
}
