% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_conclusions.R
\name{summarise_conclusions}
\alias{summarise_conclusions}
\title{Summarise counts of qualitative conclusions across all datasets}
\usage{
summarise_conclusions(ManyEcoEvo_results, ManyEcoEvo_yi_results, ManyEcoEvo)
}
\arguments{
\item{ManyEcoEvo_results}{A tibble of \code{ManyEcoEvo_results}}

\item{ManyEcoEvo_yi_results}{A tibble of \code{ManyEcoEvo_yi_results}}

\item{ManyEcoEvo}{A tibble of \code{ManyEcoEvo}}
}
\value{
A dataframe with count values for each unique \code{Conclusion} in columns for each \code{subset} ("effects", "predictions", "all"), for each \code{dataset}.
}
\description{
Summarise counts of qualitative conclusions across all datasets
}
\examples{
summarise_conclusions(ManyEcoEvo_results,ManyEcoEvo_yi_results,ManyEcoEvo)
}
