% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_conclusions.R
\name{summarise_conclusions}
\alias{summarise_conclusions}
\title{Summarise counts of qualitative conclusions across all datasets}
\usage{
summarise_conclusions(ManyEcoEvo_results, ManyEcoEvo_yi_results, ManyEcoEvo)
}
\arguments{
\item{ManyEcoEvo_results}{A tibble of \code{ManyEcoEvo_results}}

\item{ManyEcoEvo_yi_results}{A tibble of \code{ManyEcoEvo_yi_results}}

\item{ManyEcoEvo}{A tibble of \code{ManyEcoEvo}}
}
\value{
A dataframe with count values for each unique \code{Conclusion} in columns for each \code{subset} ("effects", "predictions", "all"), for each \code{dataset}.
}
\description{
Summarise counts of qualitative conclusions across all datasets
}
\details{
Data summary is generated with \code{\link[=summarise_conclusions_data]{summarise_conclusions_data()}}.
}
\examples{
summarise_conclusions(ManyEcoEvo_results,ManyEcoEvo_yi_results,ManyEcoEvo)
}
\seealso{
Other Multi-dataset Wrapper Functions: 
\code{\link{apply_VZ_exclusions}()},
\code{\link{compute_MA_inputs}()},
\code{\link{generate_exclusion_subsets}()},
\code{\link{generate_outlier_subsets}()},
\code{\link{generate_rating_subsets}()},
\code{\link{generate_yi_subsets}()},
\code{\link{make_viz}()},
\code{\link{meta_analyse_datasets}()},
\code{\link{prepare_ManyEcoEvo_yi}()},
\code{\link{prepare_ManyEcoEvo}()},
\code{\link{prepare_response_variables_yi}()},
\code{\link{prepare_response_variables}()},
\code{\link{summarise_analysis_types}()},
\code{\link{summarise_model_composition}()},
\code{\link{summarise_reviews}()},
\code{\link{summarise_sorensen_index}()},
\code{\link{summarise_variable_counts}()}
}
\author{
Hannah S. Fraser

Elliot Gould
}
\concept{Multi-dataset Wrapper Functions}
