% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_model_composition.R
\name{summarise_model_composition}
\alias{summarise_model_composition}
\title{Summarise Model Composition}
\usage{
summarise_model_composition(
  ManyEcoEvo_results,
  ManyEcoEvo_yi_results,
  ManyEcoEvo
)
}
\arguments{
\item{ManyEcoEvo_results}{A tibble of \code{ManyEcoEvo_results}}

\item{ManyEcoEvo_yi_results}{A tibble of \code{ManyEcoEvo_yi_results}}

\item{ManyEcoEvo}{A tibble of \code{ManyEcoEvo}}
}
\value{
A dataframe
}
\description{
Calculate descriptive statistics (mean, sd, min, max) for number of fixed effects, number of random effects, number of interactions and sample size of models for each dataset across both out of sample predictions and standardized corelation coefficients.
}
\examples{
summarise_model_composition(ManyEcoEvo_results, ManyEcoEvo_yi_results, ManyEcoEvo)
}
\seealso{
Other Multi-dataset Wrapper Functions: 
\code{\link{apply_VZ_exclusions}()},
\code{\link{compute_MA_inputs}()},
\code{\link{generate_exclusion_subsets}()},
\code{\link{generate_outlier_subsets}()},
\code{\link{generate_rating_subsets}()},
\code{\link{generate_yi_subsets}()},
\code{\link{make_viz}()},
\code{\link{meta_analyse_datasets}()},
\code{\link{prepare_ManyEcoEvo_yi}()},
\code{\link{prepare_ManyEcoEvo}()},
\code{\link{prepare_response_variables_yi}()},
\code{\link{prepare_response_variables}()},
\code{\link{summarise_analysis_types}()},
\code{\link{summarise_conclusions}()},
\code{\link{summarise_variable_counts}()}
}
\author{
Hannah S. Fraser

Elliot Gould
}
\concept{Multi-dataset Wrapper Functions}
