% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_review_data.R
\name{summarise_reviews}
\alias{summarise_reviews}
\title{Summarise Peer-Reviews}
\usage{
summarise_reviews(ManyEcoEvo, ManyEcoEvo_results, ManyEcoEvo_yi_results)
}
\arguments{
\item{ManyEcoEvo}{A tibble of \code{ManyEcoEvo}}

\item{ManyEcoEvo_results}{A tibble of \code{ManyEcoEvo_results}}

\item{ManyEcoEvo_yi_results}{A tibble of \code{ManyEcoEvo_yi_results}}
}
\value{
A named list of length two of \code{summarised_data_analyses} containing a tibble of summary statistics for each outcome \code{subset} (\code{effects} or \code{predictions}) generated by \code{\link[=summarise_analyses_by_reviewer]{summarise_analyses_by_reviewer()}}, and \code{summarised_data_reviews} containing a tibble of sumary statistics for each outcome \code{subset}, for each \code{dataset} generated by \code{\link[=summarise_reviews_per_analysis]{summarise_reviews_per_analysis()}}.
}
\description{
Calculates summary statistics (\code{mean}, \code{sd}, \code{min}, \code{max}) of both the number of analyses peer-reviewed by each reviewer, and the number of reviews received by each analysis.
}
\examples{
summarise_reviews(ManyEcoEvo, ManyEcoEvo_results,ManyEcoEvo_yi_results)
}
\seealso{
Other Multi-dataset Wrapper Functions: 
\code{\link{apply_VZ_exclusions}()},
\code{\link{compute_MA_inputs}()},
\code{\link{generate_exclusion_subsets}()},
\code{\link{generate_outlier_subsets}()},
\code{\link{generate_rating_subsets}()},
\code{\link{generate_yi_subsets}()},
\code{\link{make_viz}()},
\code{\link{meta_analyse_datasets}()},
\code{\link{prepare_ManyEcoEvo}()},
\code{\link{prepare_ManyEcoEvo_yi}()},
\code{\link{prepare_response_variables}()},
\code{\link{prepare_response_variables_yi}()},
\code{\link{summarise_analysis_types}()},
\code{\link{summarise_conclusions}()},
\code{\link{summarise_model_composition}()},
\code{\link{summarise_sorensen_index}()},
\code{\link{summarise_variable_counts}()}
}
\author{
Hannah S. Fraser

Elliot Gould
}
\concept{Multi-dataset Wrapper Functions}
