% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_review_data.R
\name{summarise_reviews_per_analysis}
\alias{summarise_reviews_per_analysis}
\title{Summarise reviews per each analysis}
\usage{
summarise_reviews_per_analysis(review_data)
}
\arguments{
\item{review_data}{A dataframe containing the variables \code{dataset}, \code{response_id}, and \code{ReviewerId}.}
}
\value{
A dataframe with the variables \code{dataset}, and the \code{mean}, \code{sd}, \code{min}, \code{max} of the number of times each analysis was reviewed
}
\description{
Summarise reviews per each analysis
}
