% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_variable_counts.R
\name{summarise_variable_counts}
\alias{summarise_variable_counts}
\title{Summarise variable usage across analyses}
\usage{
summarise_variable_counts(
  ManyEcoEvo,
  ManyEcoEvo_results,
  ManyEcoEvo_yi_results,
  output = "count"
)
}
\arguments{
\item{ManyEcoEvo}{A tibble of \code{ManyEcoEvo}}

\item{ManyEcoEvo_results}{A tibble of \code{ManyEcoEvo_results}}

\item{ManyEcoEvo_yi_results}{A tibble of \code{ManyEcoEvo_yi_results}}

\item{output}{A length 1 character vector equal to \code{"count"} or \code{"aggregate"}}
}
\value{
A dataframe of count values \code{n} or summary statistic values \code{n_mean}/\code{n_sd}/\code{n_min}/\code{n_max} of counts depending on the the value supplied to the \code{output} argument.
}
\description{
Calculates either the counts of analyses in which each variable is used for a given dataset, or the mean, sd, minimum and maximum of those counts.
}
\details{
To calculate count values, supply \code{"count"} to argument \code{output}. To calculate summary statistics of counts, supply \code{"aggregate"} to argument \code{output} instead.
}
\examples{
summarise_variable_usage(ManyEcoEvo, ManyEcoEvo_results, ManyEcoEvo_yi_results, "count")

summarise_variable_usage(ManyEcoEvo, ManyEcoEvo_results, ManyEcoEvo_yi_results, "aggregate")
}
\seealso{
Other Multi-dataset Wrapper Functions: 
\code{\link{apply_VZ_exclusions}()},
\code{\link{compute_MA_inputs}()},
\code{\link{generate_exclusion_subsets}()},
\code{\link{generate_outlier_subsets}()},
\code{\link{generate_rating_subsets}()},
\code{\link{generate_yi_subsets}()},
\code{\link{make_viz}()},
\code{\link{meta_analyse_datasets}()},
\code{\link{prepare_ManyEcoEvo_yi}()},
\code{\link{prepare_ManyEcoEvo}()},
\code{\link{prepare_response_variables_yi}()},
\code{\link{prepare_response_variables}()},
\code{\link{summarise_analysis_types}()},
\code{\link{summarise_conclusions}()},
\code{\link{summarise_model_composition}()},
\code{\link{summarise_sorensen_index}()}
}
\author{
Hannah S. Fraser

Elliot Gould
}
\concept{Multi-dataset Wrapper Functions}
