% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_cox_transform.R
\name{variance_box_cox}
\alias{variance_box_cox}
\title{Calculate the variance of the Box-Cox transformed absolute deviation scores}
\usage{
variance_box_cox(folded_mu, folded_v, lambda)
}
\arguments{
\item{folded_mu}{The mean of the folded absolute deviation scores}

\item{folded_v}{The variance of the folded VZr}

\item{lambda}{The lambda value used in the Box-Cox transformation}
}
\value{
The variance of the Box-Cox transformed absolute deviation scores
}
\description{
Calculate the variance of the Box-Cox transformed absolute deviation scores
}
\seealso{
Other Box-Cox transformation: 
\code{\link{box_cox_transform}()},
\code{\link{folded_params}()}

Other Analysis-level functions: 
\code{\link{Z_VZ_preds}()},
\code{\link{apply_sorensen_calc}()},
\code{\link{back_transform_response_vars_yi}()},
\code{\link{box_cox_transform}()},
\code{\link{convert_predictions}()},
\code{\link{est_to_zr}()},
\code{\link{folded_params}()},
\code{\link{pmap_wrap}()},
\code{\link{rename_prediction_cols}()}
}
\concept{Analysis-level functions}
\concept{Box-Cox transformation}
